Private API

Aerodynamics

AeroBeams.aero_steady_kinematics! — Function

aerosteadykinematics!(element::Element,V,Ω)

Computes the steady aerodynamic kinematic variables

AeroBeams.aero_unsteady_kinematics! — Function

aerounsteadykinematics!(element::Element,Vdot,Ωdot)

Computes the unsteady aerodynamic kinematic variables

AeroBeams.nondimensional_flow_parameters! — Function

nondimensionalflowparameters!(model::Model,element::Element)

Computes the nondimensional flow parameters

AeroBeams.local_gust_velocity! — Function

localgustvelocity!(problem::Problem,model::Model,element::Element)

Computes the gust velocity in the local, deformed aerodynamic basis (basis W)

AeroBeams.flap_deflection_rates! — Function

flapdeflectionrates!(problem,element::Element)

Computes the current values of flap deflection rates

AeroBeams.aero_coefficients! — Function

aero_coefficients!(problem::Problem,element::Element,χ,δNow)

Computes the aerodynamic coefficients

AeroBeams.aero_state_matrices! — Function

aerostatematrices!(element::Element,δNow)

Computes the aerodynamic state matrices

AeroBeams.aero_loads_resultants! — Function

aeroloadsresultants!(model::Model,element::Element)

Computes the aerodynamic nodal loads resultants

AeroBeams.attached_flow_aero_coefficients! — Function

attachedflowaero_coefficients!(element::Element,δNow)

Computes the aerodynamic coefficients under attached flow conditions

AeroBeams.update_airfoil_parameters! — Function

updateairfoilparameters!(problem::Problem,element::Element)

Updates the aerodynamic parameters of the airfoil according to current nondimensional flow parameters

AeroBeams.effective_angle_of_attack! — Function

effectiveangleof_attack!(element::Element,χ,δNow)

Computes the effective (unsteady) angle of attack

AeroBeams.attached_flow_cn! — Function

attachedflowcn!(element::Element,δNow)

Computes the normal force aerodynamic coefficient for attached flow

AeroBeams.attached_flow_cm! — Function

attachedflowcm!(element::Element,δNow)

Computes the pitching moment aerodynamic coefficient at the attachment point (i.e., the beam reference line) for attached flow

AeroBeams.attached_flow_ct! — Function

attachedflowct!(element::Element,δNow)

Computes the tangential force aerodynamic coefficient for attached flow

AeroBeams.pitch_plunge_effective_normalwash — Function

pitchplungeeffective_normalwash(element::Element,χ)

Computes the effective (unsteady) pitch-plunge-induced normalwash

AeroBeams.flap_effective_normalwash — Function

flapeffectivenormalwash(element::Element,χ,δNow)

Computes the effective (unsteady) flap-induced normalwash

AeroBeams.flap_normalwash — Function

flap_normalwash(element::Element,δNow)

Computes the instantaneous (quasi-steady) flap-induced normalwash

AeroBeams.gust_effective_normalwash — Function

gusteffectivenormalwash(element::Element,χ)

Computes the effective (unsteady) gust-induced normalwash

AeroBeams.flap_normalwash_rate — Function

flapnormalwashrate(element::Element,δNow)

Computes the time rate of the instantaneous flap-induced normalwash

AeroBeams.cnαUₙTQC_rate — Function

cnαUₙTQC_rate(element::Element)

Computes the time rate of the product of cnα by UₙTQC

AeroBeams.attached_flow_state_matrices! — Function

attachedflowstate_matrices!(element::Element,δNow)

Computes the aerodynamic state matrices for the indicial and inflow solvers

AeroBeams.BLi_aero_coefficients! — Function

BLiaerocoefficients!(problem::Problem,element::Element,χ,δNow)

Computes the aerodynamic coefficients according to the modified incompressible flow Beddoes-Leishman model

AeroBeams.BLi_kinematics! — Function

BLi_kinematics!(element::Element)

Computes additional kinematics associated with the Beddoes-Leishman model

AeroBeams.BLi_nonlinear_states! — Function

BLinonlinearstates!(element::Element,χ)

Sets the nonlinear states of the Beddoes-Leishman model

AeroBeams.BLi_motion_qualifiers! — Function

BLimotionqualifiers!(problem::Problem,element::Element)

Computes motion qualifiers of the modified Beddoes-Leishman model

AeroBeams.BLi_breakpoint_angles! — Function

BLibreakpointangles!(element::Element)

Computes the breakpoint angles of the modified Beddoes-Leishman model

AeroBeams.BLi_time_delays! — Function

BLitimedelays!(element::Element)

Computes time delay variables for the modified Beddoes-Leishman model

AeroBeams.BLi_separation_points! — Function

BLiseparationpoints!(element::Element)

Computes the flow separation points of the modified Beddoes-Leishman model

AeroBeams.BLi_quasi_steady_separation_points — Function

BLiquasisteadyseparationpoints(element::Element,Uₜ,Uₙ)

Computes the quasi-steady flow separation points of the modified Beddoes-Leishman model for an element

AeroBeams.BLi_stall_time! — Function

BListalltime!(problem::Problem,element::Element)

Computes variables at the time of stall onset for the modified Beddoes-Leishman model

AeroBeams.BLi_DSV_loads! — Function

BLiDSVloads!(element::Element)

Computes DSV loads for the modified Beddoes-Leishman model

AeroBeams.BLi_cn! — Function

BLi_cn!(element::Element,δNow)

Computes normal force coefficient for the modified incompressible Beddoes-Leishman model

AeroBeams.BLi_cm! — Function

BLi_cm!(element::Element,δNow)

Computes the pitching moment aerodynamic coefficient at the attachment point (i.e., the beam reference line) for the modified incompressible Beddoes-Leishman model

AeroBeams.BLi_ct! — Function

BLi_ct!(element::Element,δNow)

Computes the tangential force aerodynamic coefficient for the modified incompressible Beddoes-Leishman model

AeroBeams.BLi_state_matrices! — Function

BListatematrices!(element::Element,δNow)

Computes the aerodynamic state matrices for the modified incompressible Beddoes-Leishman model

AeroBeams.BLo_aero_coefficients! — Function

BLoaerocoefficients!(problem::Problem,element::Element,χ,δNow)

Computes the aerodynamic coefficients according to the original Beddoes-Leishman model

AeroBeams.BLo_nonlinear_states! — Function

BLononlinearstates!(element::Element,χ)

Sets the nonlinear states of the original Beddoes-Leishman model

AeroBeams.BLo_motion_qualifiers! — Function

BLomotionqualifiers!(element::Element)

Computes motion qualifiers of the original Beddoes-Leishman model

AeroBeams.BLo_breakpoint_angle! — Function

BLobreakpointangle!(element::Element)

Computes the breakpoint angle of the original Beddoes-Leishman model

AeroBeams.BLo_separation_points! — Function

BLoseparationpoints!(element::Element)

Computes the flow separation point of the original Beddoes-Leishman model

AeroBeams.BLo_stall_time! — Function

BLostalltime!(problem::Problem,element::Element)

Computes variables at the time of stall onset for the original Beddoes-Leishman model

AeroBeams.BLo_time_delays! — Function

BLotimedelays!(element::Element)

Computes time delay variables for the original Beddoes-Leishman model

AeroBeams.BLo_update_impulsive_parameters! — Function

BLoupdateimpulsive_parameters!(element::Element)

Updates the impulsive parameters for the original Beddoes-Leishman model

AeroBeams.BLo_vortex_accumulation_rate! — Function

BLovortexaccumulation_rate!(element::Element,χ)

Compute the dynamic stall vortex accumulation rate for the original Beddoes-Leishman model

AeroBeams.BLo_cn! — Function

BLo_cn!(element::Element,χ,δNow)

Computes normal force coefficient for the original Beddoes-Leishman model

AeroBeams.BLo_cm! — Function

BLo_cm!(element::Element,χ,δNow)

Computes the pitching moment aerodynamic coefficient at the attachment point (i.e., the beam reference line) for the original Beddoes-Leishman model

AeroBeams.BLo_ct! — Function

BLo_ct!(element::Element,δNow)

Computes the tangential force aerodynamic coefficient for the original Beddoes-Leishman model

AeroBeams.BLo_state_matrices! — Function

BLostatematrices!(element::Element,δNow)

Computes the aerodynamic state matrices for the original Beddoes-Leishman model

AeroBeams.update_initial_aero_states! — Function

updateinitialaero_states!(problem::Problem)

Updates the initial aerodynamic states assuming their rates are zero

AeroProperties

AeroBeams.FlowParameters — Type

mutable struct FlowParameters

FlowParameters composite type

Fields

Re = Reynolds number

Ma = Mach number

βₚ = Prandtl-Glauert compressibility factor

βₚ² = Prandtl-Glauert compressibility factor squared

Θ = Uᵢ/b*βₚ²

AeroBeams.FlowAnglesAndRates — Type

mutable struct FlowAnglesAndRates

FlowAnglesAndRates composite type

Fields

α = quasi-steady angle of attack

β = angle of sideslip

αdot = quasi-steady angle of attack rate

αₑ = effective angle of attack at 3/4-chord

AeroBeams.FlowVelocitiesAndRates — Type

mutable struct FlowVelocitiesAndRates

FlowVelocitiesAndRates composite type

Fields

U = wind velocity vector at spar position, resolved in basis W

U∞ = norm of U

Uₛ = spanwise component of U

Uₜ = tangential component of U

Uₙ = normal component of U

Uᵢ = in-plane component of U

Ωₐ = spanwise angular velocity component, resolved in basis W

UₙMid = relative normal wind velocity component at the airfoil's 1/2-chord

UₙTQC = relative normal wind velocity component at the airfoil's 3/4-chord

Udot = wind acceleration vector at spar position, resolved in basis W

Uₜdot = tangential component of Udot

Uₙdot = normal component of U

Uᵢdot = in-plane component of U

Ωₐdot = spanwise angular velocity component, resolved in basis W

UₙdotMid = relative normal wind acceleration component at the airfoil's 1/2-chord

UₙdotTQC = = relative normal wind acceleration component at the airfoil's 3/4-chord

UₜGust = gust tangential velocity component

UₙGust = gust normal velocity component

wₑp = effective pitch-plunge-induced normal velocity component at the airfoil's 3/4-chord

AeroBeams.AeroCoefficients — Type

mutable struct AeroCoefficients

AeroCoefficients composite type

Fields

cn = normal force coefficient

cm = pitching moment coefficient about spar position

ct = tangential force coefficient

cnC = circulatory component of cn

cnI = inertial component of cn

cnP = potential flow component of cn

cmC = circulatory component of cm

cmI = inertial component of cm

cmRot = rotation-induced component of cm

cnF = separated-flow circulatory component of cn

cmF = separated-flow circulatory component of cm

ctF = separated-flow circulatory component of ct

cnV = DSV-induced component of cn

cmV = DSV-induced component of cm

ctV = DSV-induced component of ct

AeroBeams.BLiNamedStates — Type

mutable struct BLiNamedStates

BLiNamedStates composite type

Fields

αlag

f2primeN

f2primeM

f2primeT

RD

RD_stallOnsetRatio

AeroBeams.BLoNamedStates — Type

mutable struct BLoNamedStates

BLoNamedStates composite type

Fields

cnPprime

f2Prime

fPrimeM

cnVP

cnVN

cnV

AeroBeams.BLiKinematics — Type

mutable struct BLiKinematics

BLiKinematics composite type

Fields

r

qR

R

AeroBeams.BLiFlowVariables — Type

mutable struct BLiFlowVariables

BLiFlowVariables composite type

Fields

stallOnsetRatio

upstroke

S

P

T

α1N

α1M

α1T

fN

fM

fT

fPrimeN

fPrimeM

fPrimeT

Ta_SO

TfN

TfM

TfT

AeroBeams.BLoFlowVariables — Type

mutable struct BLoFlowVariables

BLoFlowVariables composite type

Fields

αlag

q

stallOnsetRatio

upstroke

α1

f

fPrime

Tf

Tv

Ka

Kq

KaM

KqM

Tᵢ

Kf

cvdotP

cvdotN

AeroBeams.BLiComplementaryVariables — Type

mutable struct BLiComplementaryVariables

BLiComplementaryVariables composite type

Fields

stallOnsetRatioPrev

αlagPrev

qRPrev

PPrev

upstrokePrev

maxStallOnsetRatio

minStallOnsetRatio

qRmax

Ts

tv0P

fDiff_tv0P

qR_tv0P

R_tv0P

RD_tv0P

upstroke_tv0P

fDiff_tv0P2

RD_tv0P2

upstroke_tv0P2

tv0N

fDiff_tv0N

qR_tv0N

R_tv0N

RD_tv0N

upstroke_tv0N

fDiff_tv0N2

RD_tv0N2

upstroke_tv0N2

lastRD_tv0

AeroBeams.BLoComplementaryVariables — Type

mutable struct BLoComplementaryVariables

BLoComplementaryVariables composite type

Fields

stallOnsetRatioPrev

tv0P

tv0N

τvP

τvN

AeroBeams.AeroVariables — Type

mutable struct AeroVariables

AeroVariables composite type

Fields

flowParameters::FlowParameters

flowAnglesAndRates::FlowAnglesAndRates

flowVelocitiesAndRates::FlowVelocitiesAndRates

aeroCoefficients::AeroCoefficients

BLiKin::BLiKinematics

BLiFlow::BLiFlowVariables

BLoFlow::BLoFlowVariables

AeroBeams.AeroProperties — Type

@with_kw mutable struct AeroProperties

AeroProperties composite type

AeroBeams.initial_F_χ_χ — Function

initialFχ_χ(solver::AeroSolver,nStates::Int64)

Computes the initial value of Fχχ (for zero relative airspeed) - theoretically, the ϵ value in the function should go to zero as the airspeed goes to zero, but numerical problems are found for the inflow solver

AeroBeams.initial_F_χ_Vdot — Function

initialFχ_Vdot(solver::AeroSolver,nStates::Int64,pitchPlungeStatesRange::Union{Nothing,UnitRange{Int64}},cnα::Number)

Computes the initial value of FχVdot (for zero relative airspeed)

AeroBeams.initial_F_χ_Ωdot — Function

initialFχ_Ωdot(solver::AeroSolver,nStates::Int64,pitchPlungeStatesRange::Union{Nothing,UnitRange{Int64}},c::Number,normSparPos::Float64,cnα::Number)

Computes the initial value of FχΩdot (for zero relative airspeed)

AeroSolver

AeroBeams.theodorsen_flap_constants — Function

function theodorsenflapconstants(normSparPos::Float64,normFlapPos::Float64)

Computes Theodorsen's flap constants

AeroBeams.inflow_arrays — Function

inflow_arrays(N::Int64)

Computes the fixed state arrays from Peters' inflow theory

Airfoil

AeroBeams.AttachedFlowParameters — Type

@with_kw mutable struct AttachedFlowParameters

AttachedFlowParameters composite type

Fields

α₀N::Number

ϵₙ::Number

ϵₘ::Number

η::Number

cd₀::Number

cdδ::Number

cm₀::Number

cmα::Number

cmδ::Number

cnα::Number

cnδ::Number

AeroBeams.BLiParameters — Type

@with_kw mutable struct BLiParameters

BLiParameters composite type

Fields

α₀N::Number

αds₀::Number

αₛₛ::Number

α1₀N::Number

α1₀M::Number

α1₀T::Number

βσ1N::Number

βσ1T::Number

βσ2N::Number

βS2Nlpr::Number

βS2Tlpr::Number

βS1Nu::Number

βS1Mu::Number

βS1Tu::Number

βS1Nd::Number

βS1Md::Number

βS1Td::Number

βS2Nu::Number

βS2Mu::Number

βS2Tu::Number

βS2Nd::Number

βS2Md::Number

βS2Td::Number

γLS::Number

δα₀::Number

δα₁::Number

ϵₙ::Number

ϵₘ::Number

η::Number

κ₀::Number

κ₁::Number

κ₂::Number

κ₃::Number

λ₁::Number

λ₂::Number

μv₂::Number

ν₁::Number

ν₂::Number

ν₃::Number

ν₄::Number

ν₅::Number

χu::Number

χd::Number

ξ::Number

ζₐ::Number

cd₀::Number

cdδ::Number

cm₀::Number

cmδ::Number

cnα::Number

cnδ::Number

dt::Number

dm::Number

E₀::Number

E₁::Number

f₀N::Number

f₀M::Number

f₀T::Number

fbN::Number

fbM::Number

fbT::Number

gᵥ::Number

gᵥ₂::Number

K₀::Number

K₁::Number

K₂::Number

r₀::Number

S1N::Number

S1M::Number

S1T::Number

S2N::Number

S2M::Number

S2T::Number

Ta::Number

Tf::Number

Tg::Number

Tv::Number

Tv₂::Number

Vn₁::Number

Vn₂::Number

Vn₃::Number

Vm::Number

Vt::Number

ztd::Number

ztu::Number

zm::Number

λbWMat::Matrix{Float64}

AeroBeams.BLoParameters — Type

@with_kw mutable struct BLoParameters

BLoParameters composite type

Fields

α₀N::Number

α1₀::Number

δα::Number

ϵₙ::Number

ϵₘ::Number

η::Number

cd₀::Number

cdδ::Number

cm₀::Number

cmδ::Number

cn₁::Number

cnα::Number

cnδ::Number

Df::Number

E₀::Number

f₀::Number

fb::Number

K₀::Number

K₁::Number

K₂::Number

S1::Number

S2::Number

Tf₀::Number

Tp::Number

Tv₀::Number

TvL::Number

AeroBeams.get_airfoil_coordinates — Function

getairfoilcoordinates(name::String)

Gets the coordinates of the airfoil

Atmosphere

AeroBeams.air_properties_from_pressure_and_temperature — Function

airpropertiesfrompressureand_temperature(p::Number,T::Number,R::Number,γ::Number,β::Number,S::Number)

Calculates air properties from given pressure and temperature

Boundary conditions

AeroBeams.update_BC_data! — Function

updateBCdata!(bc::BC,timeNow::Number=0)

Updates the boundary conditions at the current time

Beam

AeroBeams.validate_beam! — Function

validate_beam!(beam::Beam)

Validates the beam inputs

AeroBeams.validate_sectional_matrices — Function

validatesectionalmatrices(beam::Beam)

Checks that sectional matrices are input as a single one for the whole beam or are input in a per element basis

AeroBeams.validate_rotation_parametrization — Function

validaterotationparametrization(beam::Beam)

Validates the input rotation parameters and rotation parametrization

AeroBeams.validate_connected_beams — Function

validateconnectedbeams(beam::Beam)

Checks that the beam connections are consistent

AeroBeams.validate_initial_conditions! — Function

validateinitialconditions!(beam::Beam)

Validate initial generalized displacements/velocities defined in basis b, and transform to function if input as constant vector

AeroBeams.validate_normalized_nodal_positions! — Function

validatenormalizednodal_positions!(beam::Beam)

Validates the normalized nodal positions if they were input, or updates them if they were not

AeroBeams.validate_hinged_nodes! — Function

validatehingednodes!(beam::Beam)

Validates the hinged nodes and DoFs

AeroBeams.validate_distributed_loads! — Function

validatedistributedloads!(beam::Beam)

Validates and updates the distributed loads

AeroBeams.velocity_dofs_to_update! — Function

velocitydofsto_update!(beam::Beam)

Gets the velocity DOFs (V and Ω) to be updated on the initial dynamic solution

AeroBeams.get_rotation_tensor! — Function

getrotationtensor!(beam::Beam)

Validates the input rotation parameters and rotation parametrization, and gets the corresponding rotation tensor from basis A to basis b

AeroBeams.initial_displacements_derivatives! — Function

initialdisplacementsderivatives!(beam::Beam)

Computes the derivatives of the initial generalized displacements with respect to the beam arclength coordinate, x₁

AeroBeams.create_beam_elements! — Function

createbeamelements!(beam::Beam)

Initializes the element and node ranges, and creates beam elements

AeroBeams.set_nodal_coordinates! — Function

setnodalcoordinates!(beam::Beam)

Sets the nodal coordinates of each element into the beam

Core

AeroBeams.element_arrays! — Function

element_arrays!(problem::Problem,model::Model,element::Element)

Computes the elemental contributions to the system's arrays (residual, Jacobian, inertia)

AeroBeams.special_node_arrays! — Function

specialnodearrays!(problem::Problem,model::Model,specialNode::SpecialNode)

Computes the nodal contributions to the system's arrays (residual, Jacobian, inertia)

AeroBeams.element_velocities_basis_b! — Function

elementvelocitiesbasis_b!(model::Model,element::Element,σ::Float64=1.0,timeNow::Number=0.0)

Computes the generalized velocities of basis b at the element's midpoint, resolved in basis A

AeroBeams.element_accelerations_basis_b! — Function

elementaccelerationsbasis_b!(model::Model,element::Element,σ::Float64=1.0,timeNow::Number=0.0)

Computes the generalized accelerations of basis b at the element's midpoint, resolved in basis A

AeroBeams.element_states! — Function

element_states!(problem::Problem,model::Model,element::Element)

Gets the states (generalized displacements, forces, velocities and aerodynamic) of the element

AeroBeams.element_states_rates! — Function

elementstatesrates!(problem::Problem,element::Element)

Computes the states' rates of the current element

AeroBeams.element_rotation_variables! — Function

elementrotationvariables!(problem::Problem,element::Element)

Computes the rotation variables for the current element

AeroBeams.element_distributed_loads! — Function

elementdistributedloads!(problem::Problem,model::Model,element::Element)

Computes the nodal resultants from distributed loads on the current element, resolved in basis A

AeroBeams.gravitational_loads! — Function

gravitational_loads!(model::Model,element::Element,σ::Float64)

Computes the nodal resultants from the distributed gravitational loads on the current element

AeroBeams.distributed_external_loads! — Function

distributedexternalloads!(problem::Problem,element::Element,σ::Float64)

Computes the nodal resultants from the externally applied distributed loads on the current element

AeroBeams.aerodynamic_loads! — Function

aerodynamic_loads!(problem::Problem,model::Model,element::Element)

Computes the nodal resultants from the aerodynamic loads on the current element

AeroBeams.wrapper_aerodynamic_loads_from_states! — Function

wrapperaerodynamicloadsfromstates!(states,problem::Problem,model::Model,element::Element)

Wrapper function for aerodynamic loads using elemental states as inputs

AeroBeams.wrapper_aerodynamic_loads_from_states_rates! — Function

wrapperaerodynamicloadsfromstates_rates!(statesRates,problem::Problem,model::Model,element::Element)

Wrapper function for aerodynamic loads using elemental states' rates as inputs

AeroBeams.aero_loads_core! — Function

aeroloadscore!(problem::Problem,model::Model,element::Element,V,Ω,χ,Vdot,Ωdot,δNow)

Computes the nodal resutants from aerodynamic loads and aerodynamic state matrices

AeroBeams.interpolate_distributed_loads — Function

interpolatedistributedloads(problem::Problem,loadArray::Array{Float64})

Interpolates the loads array at the current time

AeroBeams.element_strains! — Function

element_strains!(element::Element)

Computes the strains for the current element, resolved in basis B

AeroBeams.element_momenta! — Function

element_momenta!(element::Element)

Computes the momenta for the current element, resolved in basis B

AeroBeams.element_momenta_rates! — Function

elementmomentarates!(element::Element)

Computes the momenta rates for the current element, resolved in basis B

AeroBeams.element_residual! — Function

element_residual!(problem::Problem,model::Model,element::Element)

Computes the contributions from the current element to the residual array

AeroBeams.distributed_loads_derivatives_rotation_parameters! — Function

distributedloadsderivativesrotationparameters!(element::Element)

Computes the derivatives of the distributed loads w.r.t the rotation parameters

AeroBeams.gravitational_loads_derivatives_rotation_parameters — Function

gravitationalloadsderivativesrotationparameters(element::Element)

Computes the contributions of the gravitational loads to the Jacobian matrix

AeroBeams.distributed_external_loads_derivatives_rotation_parameters — Function

distributedexternalloadsderivativesrotation_parameters(element::Element)

Computes the contributions of the externally applied distributed loads to the Jacobian matrix

AeroBeams.aero_loads_derivatives_rotation_parameters — Function

aeroloadsderivativesrotationparameters(element::Element)

Computes the derivatives of the aerodynamic loads w.r.t the extended rotation parameters

AeroBeams.aero_derivatives! — Function

aero_derivatives!(problem::Problem,model::Model,element::Element)

Computes the derivatives of the aerodynamic loads and aerodynamic states w.r.t remaining elemental states and their rates (excludes rotation parameters, already computed)

AeroBeams.element_jacobian! — Function

element_jacobian!(problem::Problem,model::Model,element::Element)

Computes the contributions from the current element to the Jacobian matrix

AeroBeams.element_inertia! — Function

element_inertia!(problem::Problem,model::Model,element::Element)

Computes the contributions from the current element to the inertia matrix

AeroBeams.element_nodal_states! — Function

elementnodalstates!(element::Element)

Updates the nodal states of the element

AeroBeams.special_node_states! — Function

specialnodestates!(problem::Problem,model::Model,specialNode::SpecialNode)

Computes the nodal states

AeroBeams.spring_loads! — Function

spring_loads!(model::Model,specialNode::SpecialNode)

Adds the contributions of attached springs to the special node's loads

AeroBeams.special_node_residual! — Function

specialnoderesidual!(problem::Problem,model::Model,specialNode::SpecialNode)

Computes the contributions from the current node to the residual array

AeroBeams.special_node_follower_loads_derivatives_rotation_parameters! — Function

specialnodefollowerloadsderivativesrotationparameters!(problem::Problem,specialNode::SpecialNode)

Computes the contributions from the nodal follower loads to the Jacobian matrix

AeroBeams.special_node_jacobian! — Function

specialnodejacobian!(problem::Problem,model::Model,specialNode::SpecialNode)

Computes the contributions from the current node to the Jacobian matrix

AeroBeams.update_special_node_jacobian! — Function

updatespecialnodejacobian!(jacobian,forceScaling,Fp,Mp,ζonElements,eqsFu,eqsFp,eqsFF,eqsFM,eqsFFsep,eqsFMsep,DOFuF,DOFpM,DOFtrimLoads,uIsPrescribed,pIsPrescribed,isLoad=trues(6),isTrim=falses(6))

Updates the Jacobian matrix with the contributions from the current BC (if any) on the current node

AeroBeams.spring_loads_jacobians! — Function

springloadsjacobians!(model,jacobian,forceScaling,globalID,eqsFu,eqsFp,DOFuF,DOFpM,springs)

Adds the contributions of the spring loads to the Jacobian matrix

AeroBeams.element_modal_states — Function

elementmodalstates(element::Element,eigenvector::Vector{T},forceScaling::Float64)

Computes the modal states (generalized displacements, forces, strains, velocities and momenta) of the element at midpoint and at the nodes

AeroBeams.update_states! — Function

update_states!(problem::Problem)

Updates the elemental and nodal states

AeroBeams.reset_dual_numbers — Function

resetdualnumbers(obj)

Resets dual numbers from the ForwardDiff package back into their values

AeroBeams.convert_to_values — Function

converttovalues(arr)

Functional unit of resetdualnumbers()

Element

AeroBeams.ElementalStates — Type

mutable struct ElementalStates

ElementalStates composite type

Fields

u::Vector{T} = displacements resolved in basis A

p::Vector{T} = rotation parameters resolved in basis A

F::Vector{T} = sectional forces resolved in basis B

M::Vector{T} = sectional moments resolved in basis B

V::Vector{T} = sectional velocities resolved in basis B

Ω::Vector{T} = sectional angular velocities resolved in basis B

χ::Vector{T} = aerodynamic states

AeroBeams.ComplementaryElementalStates — Type

mutable struct ComplementaryElementalStates

ComplementaryElementalStates composite type

Fields

γ::Vector{T} = force strains resolved in basis b

κ::Vector{T} = moment strains resolved in basis b

P::Vector{T} = linear momenta resolved in basis B

H::Vector{T} = angular momenta resolved in basis B

AeroBeams.ElementalStatesRates — Type

mutable struct ElementalStatesRates

ElementalStatesRates composite type

Fields

udot::Vector{Float64} = displacements derivatives

pdot::Vector{Float64} = rotation parameters derivatives

Vdot::Vector{Float64} = sectional velocities derivatives

Ωdot::Vector{Float64} = sectional angular velocities derivatives

χdot::Vector{Float64} = aerodynamic states derivatives

AeroBeams.ComplementaryElementalStatesRates — Type

mutable struct ComplementaryElementalStatesRates

ComplementaryElementalStatesRates composite type

Fields

Pdot::Vector{Float64} = linear momenta derivatives

Hdot::Vector{Float64} = angular momenta derivatives

AeroBeams.NodalStates — Type

mutable struct NodalStates

NodalStates composite type

Fields

u_n1::Vector{T} = displacements of node 1 resolved in basis A

u_n2::Vector{T} = displacements of node 2 resolved in basis A

p_n1::Vector{T} = rotation parameters of node 1 resolved in basis A

p_n2::Vector{T} = rotation parameters of node 2 resolved in basis A

u_n1_b::Vector{T} = displacements of node 1 resolved in basis b

u_n2_b::Vector{T} = displacements of node 2 resolved in basis b

p_n1_b::Vector{T} = rotation parameters of node 1 resolved in basis b

p_n2_b::Vector{T} = rotation parameters of node 2 resolved in basis b

F_n1::Vector{T} = sectional forces of node 1 resolved in basis B

F_n2::Vector{T} = sectional forces of node 2 resolved in basis B

M_n1::Vector{T} = sectional moments of node 1 resolved in basis B

M_n2::Vector{T} = sectional moments of node 2 resolved in basis B

θ_n1::T = rotation angle of node 1

θ_n2::T = rotation angle of node 2

AeroBeams.Element — Type

mutable struct Element <: BeamElement

Element composite type

Notes

Finite elements belong to a Beam

AeroBeams.get_hinged_nodes_matrices — Function

gethingednodes_matrices(parent::Beam,nodesLocalID::Vector{Int64})

Gets the TF matrices resulting from hinged and not hinged nodal DoFs times the identity matrix

AeroBeams.get_element_distributed_loads — Function

getelementdistributedloads(parent::Beam,Δℓ::Number,x1n1::Number)

Gets the distributed loads in the element's local coordinate

AeroBeams.update_element_distributed_loads! — Function

updateelementdistributed_loads!(element::Element,loadType::String,loadFun::Function)

Update the distributed loads in the element's local coordinate

AeroBeams.add_point_inertia_to_element! — Function

addpointinertiatoelement!(element::Element,pointInertia::PointInertia)

Adds the point inertia's matrix to the element's sectional inertia matrix

Arguments

element::Element

pointInertia::PointInertia

Gust

AeroBeams.SharpEdgedGust — Type

@with_kw mutable struct SharpEdgedGust <: Gust

SharpEdgedGust composite type

AeroBeams.OneMinusCosineGust — Type

@with_kw mutable struct OneMinusCosineGust <: Gust

OneMinusCosineGust composite type

AeroBeams.Continuous1DGust — Type

@with_kw mutable struct Continuous1DGust <: Gust

Continuous1DGust composite type

AeroBeams.DiscreteSpaceGust — Type

@with_kw mutable struct DiscreteSpaceGust <: Gust

DiscreteSpaceGust composite type

AeroBeams.Continuous1DSpaceGust — Type

@with_kw mutable struct Continuous1DSpaceGust <: Gust

Continuous1DSpaceGust composite type

AeroBeams.Continuous2DSpaceGust — Type

@with_kw mutable struct Continuous2DSpaceGust <: Gust

ContinuousSpaceGust composite type

AeroBeams.stochastic_gust_velocity_from_white_noise — Function

stochasticgustvelocityfromwhite_noise(spectrum::String,t::Vector{<:Number},U::Number,L::Number=762,σ::Number=1)

Computes a stochastic gust velocity array by coloring a white noise signal with the appropriate spectrum

Links

AeroBeams.TrimLoadsLink — Type

@with_kw mutable struct TrimLoadsLink

TrimLoadsLink composite type

AeroBeams.FlapLink — Type

@with_kw mutable struct FlapLink

FlapLink composite type

Model

AeroBeams.Model — Type

@with_kw mutable struct Model

Model composite type

AeroBeams.validate_model! — Function

validate_model!(model::Model)

Validates the inputs to the model

AeroBeams.validate_and_update_motion_basis_A! — Function

validateandupdatemotionbasis_A!(model::Model)

Validates and updates the motion variables of basis A

AeroBeams.assemble_model! — Function

assemble_model!(model::Model,beams::Vector{Beam})

Loads the assembly of beams into the model

AeroBeams.inertia_properties! — Function

inertia_properties!(model::Model)

Computes the inertia properties of the undeformed assembly

AeroBeams.set_atmosphere! — Function

set_atmosphere!(model::Model)

Sets the atmosphere data

AeroBeams.update_linked_flap_deflections! — Function

updatelinkedflap_deflections!(model::Model)

Updates the linked flap deflections for the slave beams

AeroBeams.update_number_gust_states! — Function

updatenumbergust_states!(model::Model)

Updates the number of gust states in every element with aerodynamic surface

AeroBeams.update_loads_trim_links_global_ids! — Function

updateloadstrimlinksglobal_ids!(model::Model)

Updates the nodes' global IDs of the loads trim links

AeroBeams.update_spring_nodes_ids! — Function

updatespringnodes_ids!(model::Model)

Updates the global IDs of the springs' nodes, the list of springed nodes and the list of springs on nodes

AeroBeams.update_relative_rotation_constraint_elements_ids! — Function

updaterelativerotationconstraintelements_ids!(model::Model)

Updates the global IDs of the elements with relative rotation constraints

AeroBeams.initialize_basis_A_rotation! — Function

initializebasisA_rotation!(model::Model)

Initializes the rotation tensor from basis I (fixed, inertial) to basis A, and its transpose

AeroBeams.update_initial_conditions! — Function

updateinitialconditions!(model::Model)

Updates the initial condition states on all elements of the assembly

AeroBeams.set_BCs! — Function

set_BCs!(model::Model)

Sets the BCs onto the model

AeroBeams.get_special_nodes! — Function

getspecialnodes!(model::Model)

Gets the special nodes in the system of equations: connection, boundary, and BC'ed nodes

AeroBeams.get_system_indices! — Function

getsystemindices!(model::Model)

Gets the indices (for equations and DOFs) of the system of equations

AeroBeams.update_relative_rotation_constraint_data! — Function

updaterelativerotationconstraintdata!(model::Model)

Updates the aggregate data of relative rotation constraints

Problem

AeroBeams.Problem — Type

abstract type Problem

Defines a general Problem type

AeroBeams.ModeShape — Type

mutable struct ModeShape{T<:Union{Float64,ComplexF64}}

ModeShape composite type

AeroBeams.set_initial_states! — Function

setinitialstates!(problem::Problem)

Sets the initial elemental and nodal states into the states array

AeroBeams.initialize_system_arrays! — Function

initializesystemarrays!(problem::Problem)

Initializes the system arrays to the correct size

AeroBeams.precompute_distributed_loads! — Function

precomputedistributedloads!(problem::Problem)

Pre-computes the distributed loads over all elements' nodes, over every time step

AeroBeams.solve_steady! — Function

solve_steady!(problem::Problem)

Solves a steady problem (includes trim and the steady part of eigen problems)

AeroBeams.get_mode_shapes! — Function

getmodeshapes!(problem::Problem,eigenvectors::Matrix{T},frequencies::Vector{Float64},dampings::Vector{Float64})

Gets the mode shapes given by the eigenvectors

AeroBeams.fix_nodal_modal_states! — Function

fixnodalmodal_states!(problem::Problem)

Corrects the nodal modal states which may not coincide depending on which element is considered

AeroBeams.solve_dynamic! — Function

solve_dynamic!(problem::Problem)

Solves a dynamic problem

AeroBeams.initialize_time_variables! — Function

initializetimevariables!(problem::Problem)

Validates and initializes the time variables

AeroBeams.solve_initial_dynamic! — Function

solveinitialdynamic!(problem::Problem)

Gets consistent initial conditions and solves the first time step

AeroBeams.copy_initial_states! — Function

copyinitialstates!(problem::Problem,problemCopy::Problem)

Copies the initial states to the original problem

AeroBeams.update_initial_velocities! — Function

updateinitialvelocities!(problem::Problem)

Updates the velocity states by running a very small time step

AeroBeams.time_march! — Function

time_march!(problem::Problem)

Marches the dynamic problem in time

AeroBeams.adaptable_time_march! — Function

adaptabletimemarch!(problem::Problem)

Marches the dynamic problem in time using an adaptable time step

AeroBeams.copy_state — Function

copy_state(problem::Problem)

Copies the current state of the problem

AeroBeams.restore_state! — Function

restore_state!(problem::Problem)

Restores the state of the problem

AeroBeams.update_time_variables! — Function

updatetimevariables!(problem::Problem,timeIndex::Int64)

Updates the time variables (time, time step, time indices)

AeroBeams.update_basis_A_orientation! — Function

updatebasisAorientation!(problem::Problem,updateRAarray::Bool=true)

Updates the orientation of the basis A for the next time step

AeroBeams.get_equivalent_states_rates! — Function

getequivalentstates_rates!(problem::Problem)

Gets the equivalent states' rates at the begin of the current time step

AeroBeams.update_BL_complementary_variables! — Function

updateBLcomplementary_variables!(problem::Problem)

Updates the complementary variables of previous time step for the dynamic stall models

AeroBeams.BL_stall_boundary — Function

checkBLstall_boundary!(problem::Problem)

Computes the maximum stall boundary value over all beam elements in the Beddoes-Leishman model

AeroBeams.solve_time_step! — Function

solvetimestep!(problem::Problem)

Solves the current time step

AeroBeams.save_time_step_data! — Function

savetimestep_data!(problem::Problem,timeNow::Number)

Saves the solution at the current time step

Rotation constraint

AeroBeams.RotationConstraint — Type

@with_kw mutable struct RotationConstraint

RotationConstraint composite type

Sample models

AeroBeams.Pazy_nodal_positions — Function

Pazynodalpositions()

Gets the normalized nodal positions of the Pazy wing

AeroBeams.Pazy_stiffness_matrices — Function

Pazystiffnessmatrices(GAy::Number,GAz::Number)

Gets the sectional stiffness matrices of the Pazy wing

AeroBeams.Pazy_inertia_matrices — Function

Pazyinertiamatrices()

Gets the sectional inertia matrices of the Pazy wing

Special node

AeroBeams.SpecialNode — Type

SpecialNode(localID::Int64,globalID::Int64,connectedElementsGlobalIDs::Vector{Int64},connectedElements::Vector{Element},ζonElements::Vector{Int64},springs::Vector{Spring}=Vector{Spring}(),BCs::Vector{BC}=Vector{BC}())

Creates a special node

Arguments

localID::Int64 = local ID of the node

globalID::Int64 = global ID of the node

connectedElementsGlobalIDs::Vector{Int64} = global IDs of the connected elements

connectedElements::Vector{Element} = connected elements

ζonElements::Vector{Int64} = local coordinate on parent element

springs::Vector{Spring} = attached springs

BCs::Vector{BC} = associated boundary condtions

Spring

AeroBeams.Spring — Type

@with_kw mutable struct Spring

Spring composite type

System solver

AeroBeams.SystemSolver — Type

abstract type SystemSolver

Defines a general problem for the system of equations

AeroBeams.NewtonRaphson — Type

@with_kw mutable struct NewtonRaphson <: SystemSolver

Newton-Raphson system solver composite type

AeroBeams.solve_NewtonRaphson! — Function

solve_NewtonRaphson!(problem::Problem)

Solves the nonlinear system of equations at current time step

AeroBeams.assemble_system_arrays! — Function

assemblesystemarrays!(problem::Problem,x::Vector{Float64}=problem.x)

Assembles the residual vector, Jacobian and inertia matrices of the system of equations

AeroBeams.solve_linear_system! — Function

solvelinearsystem!(problem::Problem)

Solves the linear system of equations at current time step and load factor

AeroBeams.line_search — Function

line_search(x,residual,jacobian,λ=1e-8)

Performs a line search to solve the current Newton step

AeroBeams.line_search_step_size — Function

linesearchstep_size(x,p,residual,jacobian,c1=1e-4,ρ=0.5)

Updates the step size (α) of the line search

AeroBeams.save_load_factor_data! — Function

saveloadfactor_data!(problem::Problem,σ::Float64,x::Vector{Float64})

Saves the solution at the current load factor

AeroBeams.update_maximum_aero_loads! — Function

updatemaximumaero_loads!(problem::Problem)

Updates the maximum absolute values of aerodynamic loads

Units system

AeroBeams.UnitsSystem — Type

@with_kw mutable struct UnitsSystem

UnitsSystem composite type

AeroBeams.validate_units_system — Function

validateunitssystem(units::UnitsSystem)

Validates the units system

Utilities

AeroBeams.a1 — Constant

const a1 = [1.0; 0.0; 0.0]

First vector of basis A, resolved in that basis

AeroBeams.a2 — Constant

const a2 = [0.0; 1.0; 0.0]

Second vector of basis A, resolved in that basis

AeroBeams.a3 — Constant

const a3 = [0.0; 0.0; 1.0]

Third vector of basis A, resolved in that basis

AeroBeams.I3 — Constant

I3 = Matrix(1.0*LinearAlgebra.I,3,3)

3x3 identity matrix

AeroBeams.I6 — Constant

I6 = Matrix(1.0*LinearAlgebra.I,6,6)

6x6 identity matrix

AeroBeams.divide_inplace! — Function

divide_inplace!(divisor::Number, vars...)

Divides the input variables in-place

Arguments

divisor::Number = divisor

vars... = variables to be divided

AeroBeams.multiply_inplace! — Function

multiply_inplace!(divisor::Number, vars...)

Multiplies the input variables in-place

Arguments

multiplier::Number = multiplier

vars... = variables to be multiplied

AeroBeams.curvature_quantities — Function

curvature_quantities(k::Vector{<:Number})

Computes some useful functions of the curvature vector

AeroBeams.position_vector_from_curvature — Function

positionvectorfrom_curvature(R0::Matrix{Float64}, k::Vector{<:Number}, x1::Number)

Computes the position vector at an arclength value

AeroBeams.rotation_tensor_from_curvature — Function

rotationtensorfrom_curvature(R0::Matrix{Float64}, k::Vector{<:Number}, x1::Number)

Computes the rotation tensor at an arclength position

AeroBeams.rotation_parameter_scaling — Function

rotationparameterscaling(p::Vector{<:Number})

Scales the Wiener-Milenkovic rotation parameters

AeroBeams.rotation_tensor_derivatives_scaled_parameters — Function

rotationtensorderivativesscaledparameters(ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²)

Computes the derivatives of the rotation tensor with respect to the scaled rotation parameters

AeroBeams.scaling_derivatives_extended_parameters — Function

scalingderivativesextended_parameters(λ::Float64,p::Vector{<:Number},pNorm::Number)

Computes the derivatives of the scaling factor with respect to the extended rotation parameters

AeroBeams.rotation_tensor_derivatives_extended_parameters — Function

rotationtensorderivativesextendedparameters(p,pNorm,λ,ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²)

Computes the derivatives of the rotation tensor with respect to the extended rotation parameters

AeroBeams.rotation_tensor_time_derivative — Function

rotationtensortimederivative(Rps1,Rps2,Rps3,ps_p,pdot)

Computes the time derivative of the rotation tensor

AeroBeams.rotation_tensor_derivatives_time_extended_parameters — Function

rotationtensorderivativestimeextendedparameters(ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,Θps1,Θps2,Θps3,υ,υ²,υ²ps1,υ²ps2,υ²ps3,ps1dot,ps2dot,ps3dot,ps1p1,ps2p1,ps3p1,ps1p2,ps2p2,ps3p2,ps1p3,ps2p3,ps3p3)

Computes the derivatives of the time derivative of the rotation tensor with respect to the extended rotation parameters

AeroBeams.tangent_operator_transpose_inverse_WM — Function

tangentoperatortransposeinverseWM(p::Vector{<:Number})

Computes the inverse of the transpose of the tangent operator tensor according to Wiener-Milenkovic parameters

tangentoperatortransposeinverseWM(ps::Vector{Float64},ps0::Float64)

Computes the inverse of the transpose of the tangent operator tensor according to Wiener-Milenkovic parameters

AeroBeams.tangent_tensor_functions_derivatives_extended_parameters — Function

tangenttensorfunctionsderivativesextendedparameters(HT,ps1,ps2,ps3,υ²,υ²ps1,υ²ps2,υ²ps3,ps_p)

Computes the derivatives of the tangent tensor's transpose and its inverse with respect to the extended rotation parameters

AeroBeams.force_scaling — Function

force_scaling(S::Vector{Matrix{Float64}})

Computes the appropriate force scaling for the linear system of equations

AeroBeams.rotation_angle — Function

rotation_angle(p::Vector{<:Number})

Computes the rotation angle

AeroBeams.highest_in_rowcol — Function

highestinrowcol(O::Matrix{<:Number})

Finds the unrepeated columns of the n highest values in matrix O, where n is the size of O

Visualization

AeroBeams.plot_output_of_x1 — Function

plotoutputof_x1(beamGroups; x1,output,ind,units,YLabel,colorScheme=:rainbow,lw=1,legendPos=:best)

Plots output along the arclength coordinate for each beam group

AeroBeams.plot_output_of_time! — Function

plotoutputof_time!(plt; t,output,element=NaN,node=NaN,units,YLabel,lw=1,colorScheme=:rainbow,legendPos=:best)

Plots output over time

AeroBeams.get_undeformed_airfoil_coords — Function

getundeformedairfoil_coords(element::Element)

Computes the undeformed nodal airfoil coordinates

AeroBeams.plot_BCs! — Function

plot_BCs!(plt,problem::Problem,element::Element)

Plots all boundary conditions at the current time

AeroBeams.draw_BC! — Function

draw_BC!(plt,problem::Problem,element::Element)

Draws boundary condition

AeroBeams.draw_concentrated_force! — Function

drawconcentratedforce!(plt,problem::Problem,element::Element)

Draws concentrated force

AeroBeams.draw_concentrated_moment! — Function

drawconcentratedmoment!(plt,problem::Problem,element::Element)

Draws concentrated moment

AeroBeams.draw_generalized_displacement! — Function

drawgeneralizeddisplacement!(plt,problem::Problem,element::Element)

Draws generalized displacement boundary condition

AeroBeams.plot_distributed_loads! — Function

plotdistributedloads!(plt,problem::Problem,element::Element)

Plots all distributed loads at the current time

AeroBeams.draw_distributed_forces! — Function

drawdistributedforces!(element::Element)

Draws distributed forces

AeroBeams.draw_distributed_moments! — Function

drawdistributedmoments!(element::Element)

Draws distributed moments

AeroBeams.draw_aero_loads! — Function

drawaeroloads!(element::Element)

Draws aerodynamic loads

AeroBeams.draw_circular_vector! — Function

drawaeroloads!(element::Element)

Draws aerodynamic loads