Sample models

AeroBeams.Pazy_nodal_positions — Function

Pazynodalpositions()

Gets the normalized nodal positions of the Pazy wing

AeroBeams.Pazy_stiffness_matrices — Function

Pazystiffnessmatrices(GAy::Number,GAz::Number)

Gets the sectional stiffness matrices of the Pazy wing

AeroBeams.Pazy_inertia_matrices — Function

Pazyinertiamatrices()

Gets the sectional inertia matrices of the Pazy wing

Special node

AeroBeams.SpecialNode — Type

SpecialNode(localID::Int64,globalID::Int64,connectedElementsGlobalIDs::Vector{Int64},connectedElements::Vector{Element},ζonElements::Vector{Int64},springs::Vector{Spring}=Vector{Spring}(),BCs::Vector{BC}=Vector{BC}())

Creates a special node

Arguments

localID::Int64 = local ID of the node

globalID::Int64 = global ID of the node

connectedElementsGlobalIDs::Vector{Int64} = global IDs of the connected elements

connectedElements::Vector{Element} = connected elements

ζonElements::Vector{Int64} = local coordinate on parent element

springs::Vector{Spring} = attached springs

BCs::Vector{BC} = associated boundary condtions

Spring

AeroBeams.Spring — Type

@with_kw mutable struct Spring

Spring composite type

System solver

AeroBeams.SystemSolver — Type

abstract type SystemSolver

Defines a general problem for the system of equations

AeroBeams.NewtonRaphson — Type

@with_kw mutable struct NewtonRaphson <: SystemSolver

Newton-Raphson system solver composite type

AeroBeams.solve_NewtonRaphson! — Function

solve_NewtonRaphson!(problem::Problem)

Solves the nonlinear system of equations at current time step

AeroBeams.assemble_system_arrays! — Function

assemblesystemarrays!(problem::Problem,x::Vector{Float64}=problem.x)

Assembles the residual vector, Jacobian and inertia matrices of the system of equations

AeroBeams.solve_linear_system! — Function

solvelinearsystem!(problem::Problem)

Solves the linear system of equations at current time step and load factor

AeroBeams.line_search — Function

line_search(x,residual,jacobian,λ=1e-8)

Performs a line search to solve the current Newton step

AeroBeams.line_search_step_size — Function

linesearchstep_size(x,p,residual,jacobian,c1=1e-4,ρ=0.5)

Updates the step size (α) of the line search

AeroBeams.save_load_factor_data! — Function

saveloadfactor_data!(problem::Problem,σ::Float64,x::Vector{Float64})

Saves the solution at the current load factor

AeroBeams.update_maximum_aero_loads! — Function

updatemaximumaero_loads!(problem::Problem)

Updates the maximum absolute values of aerodynamic loads

Units system

AeroBeams.UnitsSystem — Type

@with_kw mutable struct UnitsSystem

UnitsSystem composite type

AeroBeams.validate_units_system — Function

validateunitssystem(units::UnitsSystem)

Validates the units system

Utilities

AeroBeams.a1 — Constant

const a1 = [1.0; 0.0; 0.0]

First vector of basis A, resolved in that basis

AeroBeams.a2 — Constant

const a2 = [0.0; 1.0; 0.0]

Second vector of basis A, resolved in that basis

AeroBeams.a3 — Constant

const a3 = [0.0; 0.0; 1.0]

Third vector of basis A, resolved in that basis

AeroBeams.I3 — Constant

I3 = Matrix(1.0*LinearAlgebra.I,3,3)

3x3 identity matrix

AeroBeams.I6 — Constant

I6 = Matrix(1.0*LinearAlgebra.I,6,6)

6x6 identity matrix

AeroBeams.divide_inplace! — Function

divide_inplace!(divisor::Number, vars...)

Divides the input variables in-place

Arguments

divisor::Number = divisor

vars... = variables to be divided

AeroBeams.multiply_inplace! — Function

multiply_inplace!(divisor::Number, vars...)

Multiplies the input variables in-place

Arguments

multiplier::Number = multiplier

vars... = variables to be multiplied

AeroBeams.curvature_quantities — Function

curvature_quantities(k::Vector{<:Number})

Computes some useful functions of the curvature vector

AeroBeams.position_vector_from_curvature — Function

positionvectorfrom_curvature(R0::Matrix{Float64}, k::Vector{<:Number}, x1::Number)

Computes the position vector at an arclength value

AeroBeams.rotation_tensor_from_curvature — Function

rotationtensorfrom_curvature(R0::Matrix{Float64}, k::Vector{<:Number}, x1::Number)

Computes the rotation tensor at an arclength position

AeroBeams.rotation_parameter_scaling — Function

rotationparameterscaling(p::Vector{<:Number})

Scales the Wiener-Milenkovic rotation parameters

AeroBeams.rotation_tensor_derivatives_scaled_parameters — Function

rotationtensorderivativesscaledparameters(ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²)

Computes the derivatives of the rotation tensor with respect to the scaled rotation parameters

AeroBeams.scaling_derivatives_extended_parameters — Function

scalingderivativesextended_parameters(λ::Float64,p::Vector{<:Number},pNorm::Number)

Computes the derivatives of the scaling factor with respect to the extended rotation parameters

AeroBeams.rotation_tensor_derivatives_extended_parameters — Function

rotationtensorderivativesextendedparameters(p,pNorm,λ,ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²)

Computes the derivatives of the rotation tensor with respect to the extended rotation parameters

AeroBeams.rotation_tensor_time_derivative — Function

rotationtensortimederivative(Rps1,Rps2,Rps3,ps_p,pdot)

Computes the time derivative of the rotation tensor

AeroBeams.rotation_tensor_derivatives_time_extended_parameters — Function

rotationtensorderivativestimeextendedparameters(ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,Θps1,Θps2,Θps3,υ,υ²,υ²ps1,υ²ps2,υ²ps3,ps1dot,ps2dot,ps3dot,ps1p1,ps2p1,ps3p1,ps1p2,ps2p2,ps3p2,ps1p3,ps2p3,ps3p3)

Computes the derivatives of the time derivative of the rotation tensor with respect to the extended rotation parameters

AeroBeams.tangent_operator_transpose_inverse_WM — Function

tangentoperatortransposeinverseWM(p::Vector{<:Number})

Computes the inverse of the transpose of the tangent operator tensor according to Wiener-Milenkovic parameters

tangentoperatortransposeinverseWM(ps::Vector{Float64},ps0::Float64)

Computes the inverse of the transpose of the tangent operator tensor according to Wiener-Milenkovic parameters

AeroBeams.tangent_tensor_functions_derivatives_extended_parameters — Function

tangenttensorfunctionsderivativesextendedparameters(HT,ps1,ps2,ps3,υ²,υ²ps1,υ²ps2,υ²ps3,ps_p)

Computes the derivatives of the tangent tensor's transpose and its inverse with respect to the extended rotation parameters

AeroBeams.force_scaling — Function

force_scaling(S::Vector{Matrix{Float64}})

Computes the appropriate force scaling for the linear system of equations

AeroBeams.rotation_angle — Function

rotation_angle(p::Vector{<:Number})

Computes the rotation angle

AeroBeams.highest_in_rowcol — Function

highestinrowcol(O::Matrix{<:Number})

Finds the unrepeated columns of the n highest values in matrix O, where n is the size of O

Visualization

AeroBeams.plot_output_of_x1 — Function

plotoutputof_x1(beamGroups; x1,output,ind,units,YLabel,colorScheme=:rainbow,lw=1,legendPos=:best)

Plots output along the arclength coordinate for each beam group

AeroBeams.plot_output_of_time! — Function

plotoutputof_time!(plt; t,output,element=NaN,node=NaN,units,YLabel,lw=1,colorScheme=:rainbow,legendPos=:best)

Plots output over time

AeroBeams.get_undeformed_airfoil_coords — Function

getundeformedairfoil_coords(element::Element)

Computes the undeformed nodal airfoil coordinates

AeroBeams.plot_BCs! — Function

plot_BCs!(plt,problem::Problem,element::Element)

Plots all boundary conditions at the current time

AeroBeams.draw_BC! — Function

draw_BC!(plt,problem::Problem,element::Element)

Draws boundary condition

AeroBeams.draw_concentrated_force! — Function

drawconcentratedforce!(plt,problem::Problem,element::Element)

Draws concentrated force

AeroBeams.draw_concentrated_moment! — Function

drawconcentratedmoment!(plt,problem::Problem,element::Element)

Draws concentrated moment

AeroBeams.draw_generalized_displacement! — Function

drawgeneralizeddisplacement!(plt,problem::Problem,element::Element)

Draws generalized displacement boundary condition

AeroBeams.plot_distributed_loads! — Function

plotdistributedloads!(plt,problem::Problem,element::Element)

Plots all distributed loads at the current time

AeroBeams.draw_distributed_forces! — Function

drawdistributedforces!(element::Element)

Draws distributed forces

AeroBeams.draw_distributed_moments! — Function

drawdistributedmoments!(element::Element)

Draws distributed moments

AeroBeams.draw_aero_loads! — Function

drawaeroloads!(element::Element)

Draws aerodynamic loads

AeroBeams.draw_circular_vector! — Function

drawaeroloads!(element::Element)

Draws aerodynamic loads