Public API

Creating an aerodynamic derivatives solver

AeroBeams.AD — Type

struct AD <: DerivationMethod

AD (Automatic Differentiation) DerivationMethod composite type

AeroBeams.FD — Type

struct FD <: DerivationMethod

FD (Finite Differences) DerivationMethod composite type

Fields

method::FiniteDifferenceMethod

Creating an aerodynamic solver

AeroBeams.QuasiSteady — Type

struct QuasiSteady <: AeroSolver

QuasiSteady AeroSolver composite type

AeroBeams.Indicial — Type

struct Indicial <: AeroSolver

Indicial AeroSolver composite type

AeroBeams.BLi — Type

struct BLi <: AeroSolver

Incompressible modified Beddoes-Leishman AeroSolver composite type

AeroBeams.BLo — Type

struct BLo <: AeroSolver

Original Beddoes-Leishman AeroSolver composite type

AeroBeams.Inflow — Type

struct Inflow <: AeroSolver

Inflow AeroSolver composite type

AeroBeams.TableLookup — Type

struct TableLookup <: FlapAeroSolver

TableLookup FlapAeroSolver composite type

AeroBeams.ThinAirfoilTheory — Type

mutable struct ThinAirfoilTheory <: FlapAeroSolver

ThinAirfoilTheory FlapAeroSolver composite type

AeroBeams.IndicialGust — Type

struct IndicialGust <: GustAeroSolver

IndicialGust GustAeroSolver composite type

Creating an airfoil

AeroBeams.Airfoil — Type

@with_kw mutable struct Airfoil

Airfoil composite type

Fields

name::String

coordinates::Matrix{Float64}

attachedFlowParameters::AttachedFlowParameters

parametersBLi::BLiParameters

parametersBLo::BLoParameters

AeroBeams.create_Airfoil — Function

create_Airfoil(;name::String,Re::Number=0,Ma::Number=0,U::Number=0,b::Number=0)

Initializes the airfoil with the predefined name

Arguments

name::String

Re::Number

Ma::Number

U::Number

b::Number

AeroBeams.create_flapped_Airfoil — Function

createflappedAirfoil(;name::String,flapSiteID::Int64,Re::Number=0,Ma::Number=0)

Initializes the airfoil with the predefined name and flap site ID

Arguments

name::String

flapSiteID::Int64

Re::Number

Ma::Number

U::Number

b::Number

Creating an aerodynamic surface

AeroBeams.AeroSurface — Type

@with_kw mutable struct AeroSurface

AeroSurface composite type

AeroBeams.create_AeroSurface — Function

create_AeroSurface(;solver::AeroSolver=Indicial(),flapLoadsSolver::FlapAeroSolver=ThinAirfoilTheory(),gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner"),derivationMethod::DerivationMethod=AD(),airfoil::Airfoil,c::Union{<:Function,Number},Λ::Union{<:Function,Number}=0.0,normSparPos::Union{<:Function,Float64},normFlapSpan::Union{Nothing,Vector{<:Number}}=nothing,normFlapPos::Union{Nothing,Float64}=nothing,δIsTrimVariable::Bool=false,δ::Union{Nothing,<:Function,Number}=nothing,flapSiteID::Union{Nothing,Int64}=nothing,updateAirfoilParameters::Bool=true,hasTipCorrection::Bool=false,tipLossFunction::Union{Nothing,<:Function}=nothing,tipLossDecayFactor::Number=Inf64,smallAngles::Bool=false)

Creates and aerodynamic surface

Keyword arguments

-solver::AeroSolver

flapLoadsSolver::FlapAeroSolver

gustLoadsSolver::GustAeroSolver

derivationMethod::DerivationMethod

airfoil::Airfoil

c::Union{<:Function,Number}

Λ::Union{<:Function,Number}

normSparPos::Union{<:Function,Float64}

normFlapSpan::Union{Nothing,Vector{<:Number}}

normFlapPos::Union{Nothing,Float64}

δIsTrimVariable::Bool

δ::Union{Nothing,<:Function,Number}

flapSiteID::Union{Nothing,Int64}

updateAirfoilParameters::Bool

hasTipCorrection::Bool

tipLossFunction::String

tipLossDecayFactor::Float64

smallAngles::Bool

Creating an atmosphere

AeroBeams.Atmosphere — Type

@with_kw μtable struct Atmosphere

Atmosphere composite type

Fields

ρ::Number

μ::Number

a::Number

AeroBeams.standard_atmosphere — Function

standard_atmosphere(altitude::Number)

Calculates air properties from the International Standard Atmosphere (https://en.wikipedia.org/wiki/InternationalStandardAtmosphere)

Fields

altitude::Number

Creating boundary conditions

AeroBeams.BC — Type

@with_kw mutable struct BC

Boundary conditions composite type

AeroBeams.create_BC — Function

create_BC(;name::String="",beam::Beam,node::Int64,types::Vector{String},values,toBeTrimmed::Union{BitVector,Vector{Bool}}=falses(length(types)))

BC constructor

Keyword arguments

name::String = name of the BC

beam::Beam = beam at which the BC is applied

node::Int64 = node of the beam at which the BC is applied

types::Vector{String}= types of BCs applied to the node (generalized forces and displacements)

values = corresponding values of the applied BCs (constants or functions of time)

toBeTrimmed::Union{BitVector,Vector{Bool}} = TF on whether the BC is to be trimmed

Creating a beam

AeroBeams.Beam — Type

@with_kw mutable struct Beam

Beam composite type

AeroBeams.create_Beam — Function

create_Beam()

Creates a beam

Keyword Arguments

name::String

length::Number

rotationParametrization::String

p0::Vector{<:Number}

k::Vector{<:Number}

initialPosition::Vector{<:Number}

nElements::Int64

normalizedNodalPositions::Vector{Float64}

C::Vector{<:Matrix{<:Number}}

I::Vector{<:Matrix{<:Number}}

connectedBeams::Union{Nothing,Vector{Beam}}

connectedNodesThis::Vector{Int64}

connectedNodesOther::Vector{Int64}

pointInertias::Vector{PointInertia}

hingedNodes::Vector{Int64}

hingedNodesDoF::Union{Vector{Vector{Bool}},Vector{BitVector}}

u0_of_x1::Union{Vector{<:Number},<:Function,Nothing}

p0_of_x1::Union{Vector{<:Number},<:Function,Nothing}

udot0_of_x1::Union{Vector{<:Number},<:Function,Nothing}

pdot0_of_x1::Union{Vector{<:Number},<:Function,Nothing}

f_A_of_x1t::Union{Nothing,<:Function}

m_A_of_x1t::Union{Nothing,<:Function}

f_b_of_x1t::Union{Nothing,<:Function}

m_b_of_x1t::Union{Nothing,<:Function}

ff_A_of_x1t::Union{Nothing,<:Function}

mf_A_of_x1t::Union{Nothing,<:Function}

ff_b_of_x1t::Union{Nothing,<:Function}

mf_b_of_x1t::Union{Nothing,<:Function}

aeroSurface::Union{Nothing,AeroSurface}

springs::Vector{Spring}

AeroBeams.update_beam! — Function

update_beam!(beam::Beam)

Validates and updates the beam construction

Arguments

beam::Beam

AeroBeams.add_point_inertias_to_beam! — Function

addpointinertiastobeam!(;beam::Beam,inertias::Vector{PointInertia})

Adds point inertias to the beam

Arguments

beam::Beam

Keyword arguments

inertias::Vector{PointInertia}

AeroBeams.add_loads_to_beam! — Function

addloadsto_beam!(;beam::Beam,loadTypes::Vector{String},loadFuns::Vector{<:Function})

Adds loads to the beam

Arguments

beam::Beam

Keyword arguments

loadTypes::Vector{String}

loadFuns::Vector{<:Function}

AeroBeams.add_initial_displacements_and_velocities_to_beam! — Function

addinitialdisplacementsandvelocitiestobeam!(beam::Beam,conditionTypes::Vector{String},conditionFuns::Vector{<:Function})

Adds initial generalized displacements and velocities to the beam

Arguments

beam::Beam

Keyword arguments

conditionTypes::Vector{String}

conditionFuns::Vector{<:Function}

AeroBeams.add_springs_to_beam! — Function

addspringsto_beam!(; beam::Beam,springs::Vector{Spring})

Adds simply-attached springs to a beam

Keyword arguments

beam::Beam

springs::Vector{Spring}

AeroBeams.add_spring_to_beams! — Function

addspringto_beams!(; beams::Vector{Beam},spring::Spring)

Adds a doubly-attached spring to the beams

Keyword arguments

beams::Vector{Beam}

spring::Spring

Creating a gust

AeroBeams.create_SharpEdgedGust — Function

create_SharpEdgedGust(; initialTime::Number=0,duration::Number=Inf,convectiveVelocity::Number=0,verticalVelocity::Number,p::Vector{<:Number}=zeros(3))

Creates a sharp-edged gust

Keyword arguments

initialTime::Number = time when the gust begins

duration::Number = duration of the gust

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

verticalVelocity::Number = peak "vertical" velocity of the gust (in lift direction)

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

AeroBeams.create_OneMinusCosineGust — Function

create_OneMinusCosineGust(; initialTime::Number=0,duration::Number,convectiveVelocity::Number=0,verticalVelocity::Number,p::Vector{<:Number}=zeros(3))

Creates a one-minus-cosine gust

Keyword arguments

initialTime::Number = time when the gust begins

duration::Number = duration of the gust

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

verticalVelocity::Number = peak "vertical" velocity of the gust (in lift direction)

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

AeroBeams.create_Continuous1DGust — Function

create_Continuous1DGust(; spectrum::String="vK",generationMethod::String="sinusoids",initialTime::Number=0,duration::Number,generationDuration::Number=duration,L::Number=762,ωmin::Number=0,ωmax::Number=200π,Uref::Number,convectiveVelocity::Number=0,σ::Number,p::Vector{<:Number}=zeros(3),seed::Int64=123456,plotPSD::Bool=false)

Creates a continuous 1D gust

Keyword arguments

spectrum::String = spectrum of the PSD: von Kármán ("vK") or Dryden ("Dryden")

generationMethod::String = method for generation of gust velocity ("sinusoids" or "whiteNoise")

initialTime::Number = time when the gust begins

duration::Number = duration of the gust

generationDuration::Number = duration of the gust considered for its generation

L::Number = turbulence length scale [m]

ωmin::Number = minimum frequency of the PSD [rad/s]

ωmax::Number = maximum frequency of the PSD [rad/s]

Uref::Number = reference airspeed

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

σ::Number = turbulence intensity (RMS) of "vertical" gust velocity component

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

seed::Int64 = seed for random numbers generation (reproducibility)

plotPSD::Bool = TF to plot the PSD

AeroBeams.create_DiscreteSpaceGust — Function

create_DiscreteSpaceGust(; type::String,length::Number,width::Number,convectiveVelocity::Number=0,verticalVelocity::Number,c0::Vector{<:Number}=zeros(3),p::Vector{<:Number}=zeros(3))

Creates a discrete space gust

Keyword arguments

type::String = type of gust

length::Number = length of the gust (in longitudinal direction)

width::Number = width of the gust (in spanwise direction)

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

verticalVelocity::Number = peak "vertical" velocity of the gust (in lift direction)

c0::Vector{<:Number} = position vector of the front of the gust, resolved in the inertial basis

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

AeroBeams.create_Continuous1DSpaceGust — Function

create_Continuous1DSpaceGust(; spectrum::String,length::Number,N::Int64=1001,L::Number=762,σ::Number=1,c0::Vector{<:Number}=zeros(3),p::Vector{<:Number}=zeros(3),seed::Int64=123456)

Creates a continuous 1D space gust

Keyword arguments

spectrum::String = spectrum of the PSD: von Kármán ("vK") or Dryden ("Dryden")

length::Number = length of the gust (in longitudinal direction)

N::Int64 = number of nodes for length discretization

L::Number = turbulence length scale

σ::Number = turbulence intensity (RMS) of "vertical" gust velocity component

c0::Vector{<:Number} = position vector of the front of the gust, resolved in the inertial basis

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

seed::Int64 = seed for random numbers generation (reproducibility)

AeroBeams.create_Continuous2DSpaceGust — Function

create_Continuous2DSpaceGust(; spectrum::String,length::Number,width::Number,Nx::Int64=101,Ny::Int64=101,L::Number=762,σ::Number=1,c0::Vector{<:Number}=zeros(3),p::Vector{<:Number}=zeros(3),seed::Int64=123456)

Creates a continuous 2D space gust

Keyword arguments

spectrum::String = spectrum of the PSD: von Kármán ("vK") or Dryden ("Dryden")

length::Number = length of the gust (in longitudinal direction)

width::Number = width of the gust (in lateral direction)

Nx::Int64 = number of nodes for length discretization

Ny::Int64 = number of nodes for width discretization

L::Number = turbulence length scale

σ::Number = turbulence intensity (RMS) of "vertical" gust velocity component

c0::Vector{<:Number} = position vector of the front of the gust, resolved in the inertial basis

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

seed::Int64 = seed for random numbers generation (reproducibility)

Creating a link

AeroBeams.create_TrimLoadsLink — Function

create_TrimLoadsLink(;masterBC::BC,slaveBCs::Vector{BC})

Creates a link between trim loads so that they are equal

Keyword arguments

masterBC::BC = master BC

slaveBCs::Vector{BC} = slave BCs

AeroBeams.create_FlapLink — Function

create_FlapLink(;masterBeam::Beam,slaveBeams::Vector{Beam},δMultipliers::Vector{<:Number}=ones(length(slaveBeams)))

Creates a link between flapped surfaces

Keyword arguments

masterBeam::Beam = beam of the master surface

slaveBeams::Vector{Beam} = beams of the slave surfaces

δMultipliers::Vector{<:Number} = multiplication factors of flap deflection in slave surfaces relative to the master surface

Creating a model

AeroBeams.create_Model — Function

createModel(;name::String="",units::UnitsSystem=createUnitsSystem(),beams::Vector{Beam},initialPosition::Vector{<:Number}=zeros(3),gravityVector::Vector{<:Number}=zeros(3),BCs::Vector{BC}=Vector{BC}(),pA0::Vector{Float64}=zeros(3),uA::Union{Vector{<:Number},<:Function,Nothing}=nothing,vA::Union{Vector{<:Number},<:Function,Nothing}=nothing,ωA::Union{Vector{<:Number},<:Function,Nothing}=nothing,vdotA::Union{Vector{<:Number},<:Function,Nothing}=nothing,ωdotA::Union{Vector{<:Number},<:Function,Nothing}=nothing,altitude::Union{Nothing,Number}=nothing,atmosphere::Union{Nothing,Atmosphere}=nothing,gust::Union{Nothing,Gust}=nothing,trimLoadsLinks::Vector{TrimLoadsLink}=Vector{TrimLoadsLink}(),flapLinks::Vector{FlapLink}=Vector{FlapLink}(),rotationConstraints::Vector{RotationConstraint}=Vector{RotationConstraint}())

Creates a model

Keyword arguments

name::String = name of the model

units::create_UnitsSystem = units system

beams::Vector{Beam} = beams in the assembly

initialPosition::Vector{<:Number} = initial position vector of the first node of the first beam, resolved in the inertial frame I

gravityVector::Vector{<:Number} = gravity vector

BCs::Vector{BC} = boundary condtions

p_A0::Vector{Float64} = initial rotation parameters that bring basis I to basis A

u_A::Union{Vector{<:Number},<:Function,Nothing} = displacement of basis A relative to basis I

v_A::Union{Vector{<:Number},<:Function,Nothing} = velocity of basis A relative to basis I

ω_A::Union{Vector{<:Number},<:Function,Nothing} = angular velocity of basis A relative to basis I

vdot_A::Union{Vector{<:Number},<:Function,Nothing} = acceleration of basis A relative to basis I

ωdot_A::Union{Vector{<:Number},<:Function,Nothing} = angular acceleration of basis A relative to basis I

altitude::Union{Nothing,Number} = altidude

atmosphere::Union{Nothing,Atmosphere} = atmosphere

gust::Union{Nothing,Gust} = gust

trimLoadsLinks::Vector{TrimLoadsLink} = links between trim loads

flapLinks::Vector{FlapLink} = links between flapped surfaces

rotationConstraints::Vector{RotationConstraint} = rotation constraints

AeroBeams.update_model! — Function

update_model!(model::Model)

Updates the model with its current settings

Arguments

model::Model

AeroBeams.set_motion_basis_A! — Function

setmotionbasisA!(; model::Model,uA::Union{Vector{<:Number},<:Function,Nothing}=nothing,vA::Union{Vector{<:Number},<:Function,Nothing}=nothing,ωA::Union{Vector{<:Number},<:Function,Nothing}=nothing,vdotA::Union{Vector{<:Number},<:Function,Nothing}=nothing,ωdotA::Union{Vector{<:Number},<:Function,Nothing}=nothing)

Sets the motion of basis A into the model

Keyword arguments

model::Model = model

u_A::Union{Vector{<:Number},<:Function,Nothing} = displacement of basis A relative to basis I

v_A::Union{Vector{<:Number},<:Function,Nothing} = velocity of basis A relative to basis I

ω_A::Union{Vector{<:Number},<:Function,Nothing} = angular velocity of basis A relative to basis I

vdot_A::Union{Vector{<:Number},<:Function,Nothing} = acceleration of basis A relative to basis I

ωdot_A::Union{Vector{<:Number},<:Function,Nothing} = angular acceleration of basis A relative to basis I

Creating a point inertia

AeroBeams.PointInertia — Type

@with_kw mutable struct PointInertia

PointInertia composite type

Fields

elementID::Int64 = local element ID to which the point inertia is attached

mass::Number = mass

η::Vector{Number} = position relative to element's midpoint's reference line

Ixx::Number = mass moment of inertia about the x1-axis

Iyy::Number = mass moment of inertia about the x2-axis

Izz::Number = mass moment of inertia about the x3-axis

Ixy::Number = mass product of inertia with respect to the x1-axis and x2-axis

Ixz::Number = mass product of inertia with respect to the x1-axis and x3-axis

Iyz::Number = mass product of inertia with respect to the x2-axis and x3-axis

inertiaMatrix::Union{Matrix{Number},Matrix{Nothing}} = mass moment of inertia matrix (tensor)

Creating a problem

AeroBeams.InitialVelocitiesUpdateOptions — Type

@with_kw mutable struct InitialVelocitiesUpdateOptions

InitialVelocitiesUpdateOptions composite type

Defines variables for the update of the initial velocities states

Fields

Δt::Number = time step

maxIter::Int64 = maximum number of iterations

relaxFactor::Float64 = relaxation factor

tol::Float64 = convergence tolerance

displayProgress::Bool = flag to display progress

AeroBeams.SteadyProblem — Type

@with_kw mutable struct SteadyProblem <: Problem

SteadyProblem composite type

AeroBeams.create_SteadyProblem — Function

createSteadyProblem(;model::Model,systemSolver::SystemSolver=createNewtonRaphson(),getLinearSolution::Bool=false,x0::Vector{Float64}=zeros(0))

Steady problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

x0::Vector{Float64} = initial states

AeroBeams.TrimProblem — Type

@with_kw mutable struct TrimProblem <: Problem

TrimProblem composite type

AeroBeams.create_TrimProblem — Function

createTrimProblem(;model::Model,systemSolver::SystemSolver=createNewtonRaphson(),getLinearSolution::Bool=false,getInertiaMatrix::Bool=true,x0::Vector{Float64}=zeros(0))

Trim problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

getInertiaMatrix::Bool = flag to compute inertia matrix

x0::Vector{Float64} = initial states

AeroBeams.EigenProblem — Type

@with_kw mutable struct EigenProblem <: Problem

EigenProblem composite type

AeroBeams.create_EigenProblem — Function

createEigenProblem(;model::Model,systemSolver::SystemSolver=createNewtonRaphson(),getLinearSolution::Bool=false,nModes::Int64=Inf64,frequencyFilterLimits::Vector{Float64}=[0,Inf64],normalizeModeShapes::Bool=true,x0::Vector{Float64}=zeros(0),jacobian::SparseMatrixCSC{Float64,Int64}=spzeros(0,0),inertia::SparseMatrixCSC{Float64,Int64}=spzeros(0,0))

Eigen problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

nModes::Int64=Inf64 = number of modes to be computed

frequencyFilterLimits::Vector{Float64} = limits of the frequency filter

normalizeModeShapes::Bool = flag to normalize mode shapes

x0::Vector{Float64} = initial states

jacobian::SparseMatrixCSC{Float64,Int64} = Jacobian matrix

inertia::SparseMatrixCSC{Float64,Int64} = inertia matrix

AeroBeams.DynamicProblem — Type

@with_kw mutable struct DynamicProblem <: Problem

DynamicProblem composite type

AeroBeams.create_DynamicProblem — Function

createDynamicProblem(;model::Model,systemSolver::SystemSolver=createNewtonRaphson(),getLinearSolution::Bool=false,initialTime::Number=0.0,Δt::Union{Nothing,Number}=nothing,finalTime::Union{Nothing,Number}=nothing,timeVector::Union{Nothing,Vector{Float64}}=nothing,adaptableΔt::Bool=false,minΔt::Union{Nothing,Number}=nothing,maxΔt::Union{Nothing,Number}=nothing,δb::Float64=-1e-5,skipInitialStatesUpdate::Bool=false,initialVelocitiesUpdateOptions::InitialVelocitiesUpdateOptions=InitialVelocitiesUpdateOptions(),trackingTimeSteps::Bool=true,trackingFrequency::Int64=1,displayProgress::Bool=true,displayFrequency::Int64=0,x0::Vector{Float64}=zeros(0))

Dynamic problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

initialTime::Number = initial time

Δt::Union{Nothing,Number} = time step

finalTime::Union{Nothing,Number} = final time

timeVector::Union{Nothing,Vector{Float64}} = time vector

adaptableΔt::Bool = flag for adaptable time step

minΔt::Union{Nothing,Number} = minimum time step (when adaptable)

maxΔt::Union{Nothing,Number} = maximum time step (when adaptable)

δb::Float64 = discontinuities boundary convergence norm

skipInitialStatesUpdate::Bool = flag to skip update of initial states

initialVelocitiesUpdateOptions::InitialVelocitiesUpdateOptions = options for the initial velocities update

trackingTimeSteps::Bool = flag to track time step solutions

trackingFrequency::Int64 = frequency of time steps in which to track solution

displayProgress::Bool = flag to display progress

displayFrequency::Int64 = frequency of time steps in which to display progress

x0::Vector{Float64} = initial states

AeroBeams.solve! — Function

solve!(problem::Problem)

Solves a problem

Arguments

problem::Problem

AeroBeams.solve_eigen! — Function

solve_eigen!(problem::Problem)

Solves an eigenproblem

Creating a rotation constraint

AeroBeams.create_RotationConstraint — Function

create_RotationConstraint(; masterBeam::Beam,slaveBeam::Beam,masterElementLocalID::Int64,slaveElementLocalID::Int64,DOF::Int64,value::Number)

Rotation constraint constructor

Keyword arguments

masterBeam::Beam = master beam

slaveBeam::Beam = slave beam

masterElementLocalID::Int64 = local ID of the master element

slaveElementLocalID::Int64 = local ID of the slave element

DOF::Int64 = constrained rotation degree-of-freedom

value::Number = value of the slave DOF relative to the master DOF

Creating a sample model

AeroBeams.create_Pazy — Function

create_Pazy(; p0::Vector{<:Number}=zeros(3),airfoil::Airfoil=NACA0018,aeroSolver::AeroSolver=Indicial(),gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner"),derivationMethod::DerivationMethod=AD(),withTipCorrection::Bool=true,GAy::Number=1e16,GAz::Number=GAy)

Creates the Pazy wing

Returns the beam and geometrical properties

Arguments

p0::Vector{<:Number} = initial rotation parameters

airfoil::Airfoil=NACA0018 = airfoil section

aeroSolver::AeroSolver=Indicial() = aerodynamic solver

gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner") = indicial gust loads solver

derivationMethod::DerivationMethod=AD() = method for aerodynamic derivatives

withTipCorrection::Bool=true = flag for aerodynamic tip correction

GAy::Number = shear stiffness in the x2 direction

GAz::Number = shear stiffness in the x3 direction

AeroBeams.create_PazyFFWT — Function

create_PazyFFWT(; p0::Vector{<:Number}=zeros(3),airfoil::Airfoil=deepcopy(NACA0018),aeroSolver::AeroSolver=Indicial(),gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner"),derivationMethod::DerivationMethod=AD(),withTipCorrection::Bool=false,GAy::Number=1e16,GAz::Number=GAy,hingeNode::Int64=14,hingeAngle::Number=0,flareAngle::Number=10,kSpring::Number=1e6,g::Number=0,airspeed::Number)

Creates a version of the Pazy wing with flared folding wingtip (FFWT)

Arguments

p0::Vector{<:Number} = initial rotation parameters

airfoil::Airfoil=NACA0018 = airfoil section

aeroSolver::AeroSolver=Indicial() = aerodynamic solver

gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner") = indicial gust loads solver

derivationMethod::DerivationMethod=AD() = method for aerodynamic derivatives

withTipCorrection::Bool=true = flag for aerodynamic tip correction

GAy::Number = shear stiffness in the x2 direction

GAz::Number = shear stiffness in the x3 direction

hingeNode::Int64 = hinge node

hingeAngle::Number = hinge (fold) angle

flareAngle::Number = flare angle

kSpring::Number = stiffness of the hinge

g::Number=0 = local acceleration of gravity

airspeed::Number = local airspeed

AeroBeams.Pazy_tip_loss_factor — Function

Pazytiploss_factor(αᵣ::Number,U::Number)

Computes the tip loss factor for the Pazy wing's tip correction function

Arguments

αᵣ::Number = root pitch angle, in degrees

U::Number = airspeed

AeroBeams.create_Helios — Function

create_Helios(;altitude::Number=0,aeroSolver::AeroSolver=Indicial(),gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner"),derivationMethod::DerivationMethod=AD(),g::Number=-9.80665,wingAirfoil::Airfoil=HeliosWingAirfoil,podAirfoil::Airfoil=HeliosPodAirfoil,beamPods::Bool=false,stiffnessFactor::Number=1.0,∞::Number=1e12,nElemStraightSemispan::Int64=10,nElemDihedralSemispan::Int64=5,nElemPod::Int64=1,payloadPounds::Number=0,airspeed::Number=0,δIsTrimVariable::Bool=false,thrustIsTrimVariable::Bool=false,δ::Union{Nothing,Number,<:Function}=nothing,thrust::Union{Number,<:Function}=0,reducedChord::Bool,payloadOnWing::Bool=false)

Creates a model based on the flying-wing aircraft described by Patil and Hodges in: Flight Dynamics of Highly Flexible Flying Wings (2006)

Keyword arguments

altitude::Number = altitude

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

g::Number = local acceleration of gravity

wingAirfoil::Airfoil = airfoil section of the wing

podAirfoil::Airfoil = airfoil section of the pods

beamPods::Bool = flag to include pods

stiffnessFactor::Number = stiffness factor for the wing structure

∞::Number = value for rigid structural properties

nElemStraightSemispan::Int64 = number of elements in the straight section of the semispan

nElemDihedralSemispan::Int64 = number of elements in the dihedral section of the semispan

nElemPod::Int64 = number of elements in the pods

payloadPounds::Number = payload, in pounds

airspeed::Number = local initial/trim airspeed

δIsTrimVariable::Bool = flag for flap deflection being a trim variable

thrustIsTrimVariable::Bool = flag for motors' thrust being a trim variable

δ::Union{Nothing,Number,<:Function} = flap deflection

thrust::Union{Number,<:Function} = motors' thrust

reducedChord::Bool = flag to employ a reduced (7 ft) chord

payloadOnWing::Bool = flag to set the payload on the wing's reference line

AeroBeams.create_conventional_HALE — Function

createconventionalHALE(; altitude::Number=20e3,aeroSolver::AeroSolver=Indicial(),derivationMethod::DerivationMethod=AD(),flapLoadsSolver::FlapAeroSolver=TableLookup(),gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner"),stabilizersAero::Bool=true,includeVS::Bool=true,nElemWing::Int64=20,nElemHorzStabilizer::Int64=10,nElemTailBoom::Int64=10,nElemVertStabilizer::Int64=5,∞::Number=1e12,stiffnessFactor::Number=1,k1::Number=0,k2::Number=0,airspeed::Number=0,δElevIsTrimVariable::Bool=false,thrustIsTrimVariable::Bool=false,δElev::Union{Nothing,Number,<:Function}=nothing,thrust::Union{Number,<:Function}=0,g::Number=-9.80665,wingCd0::Number=0,wingcnδ::Number=2.5,wingcmδ::Number=-0.35,wingcdδ::Number=0.15,stabsCd0::Number=0,stabscnδ::Number=2.5,stabscmδ::Number=-0.35,stabscdδ::Number=0.15)

Creates a model based on the conventional HALE aircraft described by Patil, Hodges and Cesnik in: Nonlinear Aeroelasticity and Flight Dynamics of HALE (2001)

Keyword arguments

altitude::Number = altitude

aeroSolver::AeroSolver = aerodynamic solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

flapLoadsSolver::FlapAeroSolver = aerodynamic solver for flap loads

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

stabilizersAero::Bool = flag for stabilizers with aerodynamic surfaces

includeVS::Bool = flag to include a vertical stabilizer in the model

nElemWing::Int64 = number of elements of the full wing

nElemHorzStabilizer::Int64 = number of elements of the horizontal stabilizer

nElemTailBoom::Int64 = number of elements of the tail boom

nElemVertStabilizer::Int64 = number of elements of the vertical stabilizer

∞::Number=1e12 = value of rigid structural properties

stiffnessFactor::Number = stiffness factor for the wing structure

k1::Number = undeformed wing torsional curvature

k2::Number = undeformed wing flapwise bending curvature

airspeed::Number = local initial/trim airspeed

δElevIsTrimVariable::Bool = flag for elevator deflection being a trim variable

thrustIsTrimVariable::Bool = flag for motors' thrust being a trim variable

δElev::Union{Nothing,Number,<:Function} = elevator deflection

thrust::Union{Number,<:Function} = motors' thrust

g::Number = local acceleration of gravity

wingCd0::Number = parisite drag coefficient for the wing

wingcnδ::Number = cn vs δ slope for the wing

wingcmδ::Number = cm vs δ slope for the wing

wingcdδ::Number = cd vs δ slope for the wing

stabsCd0::Number = parisite drag coefficient for the stabilizers

stabscnδ::Number = cn vs δ slope for the stabilizers

stabscmδ::Number = cm vs δ slope for the stabilizers

stabscdδ::Number = cd vs δ slope for the stabilizers

AeroBeams.create_BWB — Function

create_BWB(; altitude::Number=0,aeroSolver::AeroSolver=Indicial(),gustLoadsSolver::GustAeroSolver=IndicialGust("Kussner"),derivationMethod::DerivationMethod=AD(),∞::Number=1e12,stiffnessFactor::Number=1,airspeed::Number=0,δElevIsTrimVariable::Bool=false,thrustIsTrimVariable::Bool=false,δElev::Union{Nothing,Number,<:Function}=nothing,thrust::Union{Number,<:Function}=0,g::Number=-9.80665,updateAirfoilParameters::Bool=false,hasTipCorrection::Bool=false,tipLossDecayFactor::Number=40)

Creates a model based on the blended-wing-body described by Weihua Su's PhD thesis

Keyword arguments

altitude::Number = altitude

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

∞::Number=1e12 = value of rigid structural properties

stiffnessFactor::Number = stiffness factor for the wing structure

airspeed::Number = local initial/trim airspeed

δElevIsTrimVariable::Bool = flag for elevator deflection being a trim variable

thrustIsTrimVariable::Bool = flag for motors' thrust being a trim variable

δElev::Union{Nothing,Number,<:Function} = elevator deflection

thrust::Union{Number,<:Function} = motors' thrust

g::Number = local acceleration of gravity

updateAirfoilParameters::Bool = flag to update airfoil parameters with airspeed

hasTipCorrection::Bool = flag to employ aerodynamic tip correction

tipLossDecayFactor::Number = tip loss decay factor

Creating a spring

AeroBeams.create_Spring — Function

create_Spring(; basis::String="A",elementsIDs::Vector{Int64},nodesSides::Vector{Int64},ku::Vector{<:Number}=zeros(3),kp::Vector{<:Number}=zeros(3),kTranslational::Number=0,kTwist::Number=0,kIPBending::Number=0,kOOPBending::Number=0)

Creates a spring

Keyword arguments

basis::String = basis on which stiffnesses are defined

elementsIDs::Vector{Int64} = local IDs of the element(s)' node(s) to which the spring is attached

nodesSides::Vector{Int64} = sides (1 or 2) of the node(s) to which the spring is attached

ku::Vector{<:Number} = translational stiffness vector

kp::Vector{<:Number} = rotational stiffness vector

kTranslational::Number = translational stiffness (same in all directions)

kTwist::Number = twist stiffness (rotation about x1 direction)

kIPBending::Number = in-plane bending stiffness (rotation about x3 direction)

kOOPBending::Number = out-of-plane bending stiffness (rotation about x2 direction)

Creating a system solver

AeroBeams.create_NewtonRaphson — Function

create_NewtonRaphson(; absoluteTolerance::Float64=1e-8,relativeTolerance::Float64=1e-8,maximumIterations::Int64=20,desiredIterations::Int64=5,maximumAbsoluteError::Number=1e6,maximumRelativeError::Number=1e6,initialLoadFactor::Number=1.0,minimumLoadFactor::Float64=0.01,maximumLoadFactorStep::Float64=0.5,minimumLoadFactorStep::Float64=0.01,ρ::Float64=1.0,trackingLoadSteps::Bool=true,displayStatus::Bool=false,minConvRateAeroJacUpdate::Number=2.0,minConvRateJacUpdate::Number=2.0,alwaysUpdateJacobian::Bool=true)

Newton-Raphson nonlinear system solver constructor

Keyword arguments

absoluteTolerance::Float64 = absolute convergence tolerance

relativeTolerance::Float64 = relative convergence tolerance

maximumIterations::Int64 = maximum number of iterations

desiredIterations::Int64 = desired number of iterations

maximumAbsoluteError::Number = maximum absolute error for divergence detection

maximumRelativeError::Number = maximum relative error for divergence detection

initialLoadFactor::Number = initial load factor

minimumLoadFactor::Float64 = minimum load factor

maximumLoadFactorStep::Float64 = maximum load factor step

minimumLoadFactorStep::Float64 = minimum load factor step

ρ::Float64 = relaxation factor for trim variables

trackingLoadSteps::Bool = flag to track partial load steps solutions

displayStatus::Bool = flag to display status

minConvRateAeroJacUpdate::Number = minimum convergence rate to skip computation of aerodynamic Jacobians

minConvRateJacUpdate::Number = minimum convergence rate to skip computation of structural Jacobians

alwaysUpdateJacobian::Bool = flag to update Jacobians on every iteration

Creating a units system

AeroBeams.create_UnitsSystem — Function

create_UnitsSystem(;length::String="m",force::String="N",angle::String="rad",frequency::String="rad/s",mass::String="kg")

Creates a system composed of length, force, angle, frequency and mass units (this is only for plotting purposes and does not influence calculations)

Keyword arguments

length::String = length unit

force::String = force unit

angle::String = angle unit

frequency::String = frequency unit

mass::String = mass unit

Utilities

AeroBeams.round_off! — Function

round_off!(x)

Rounds the array or number to input tolerance (defaults to machine epsilon)

Arguments

x = array / number

tol = tolerance

AeroBeams.rms — Function

rms(x)

Computes the root mean square of an array

AeroBeams.tilde — Function

tilde(v::Vector{<:Number})

Computes the skew-symmetric matrix associated with a vector

Arguments

v::Vector{<:Number} = three-element vector

AeroBeams.mul3 — Function

mul3(A1,A2,A3,b)

Computes the scalar product of a third-order tensor represented by matrices A1, A2, and A3 with the vector b

AeroBeams.isotropic_stiffness_matrix — Function

isotropicstiffnessmatrix(;EA::Number,GAy::Number,GAz::Number,GJ::Number,EIy::Number,EIz::Number)

Creates a 6x6 sectional stiffness matrix

Arguments

∞::Number = value for rigid properties

EA::Number = axial stiffness

GAy::Number = shear stiffness in the x2 direction

GAz::Number = shear stiffness in the x3 direction

GJ::Number = torsional stiffness

EIy::Number = bending stiffness about the x2 direction

EIz::Number = bending stiffness about the x3 direction

AeroBeams.inertia_matrix — Function

inertia_matrix(;ρA::Number,ρIy::Number=0,ρIz::Number=0,ρIs::Number=ρIy+ρIz,e2::Number=0,e3::Number=0)

Creates a 6x6 sectional inertia matrix

Arguments

ρA::Number = mass per unit length

ρIy::Number = mass moment of inertia per unit length about the x2-axis

ρIz::Number = mass moment of inertia per unit length about the x3-axis

ρIs::Number = mass moment of inertia per unit length about the x1-axis

e2::Number = offset of center of gravity in the x2 direction with respect to the local reference line

e3::Number = offset of center of gravity in the x3 direction with respect to the local reference line

AeroBeams.rotation_tensor_E321 — Function

rotationtensorE321(p::Vector{<:Number})

Computes the rotation tensor according to Euler parameters sequence 3-2-1

Arguments

p::Vector{<:Number} = rotation parameters

AeroBeams.rotation_tensor_E313 — Function

rotationtensorE321(p::Vector{<:Number})

Computes the rotation tensor according to Euler parameters sequence 3-1-3

Arguments

p::Vector{<:Number} = rotation parameters

AeroBeams.rotation_tensor_WM — Function

rotationtensorWM(p::Vector{<:Number})

Computes the rotation tensor according to Wiener-Milenkovic parameters

Arguments

p::Vector{<:Number} = rotation parameters

AeroBeams.tangent_operator_transpose_WM — Function

tangentoperatortranspose_WM(p::Vector{Float64})

Computes the transpose of tangent operator tensor according to Wiener-Milenkovic parameters

Arguments

p::Vector{Float64} = rotation parameters

tangentoperatortranspose_WM(ps::Vector{Float64},ps0::Float64,υ²::Float64)

Computes the transpose of tangent operator tensor according to Wiener-Milenkovic parameters

AeroBeams.tangent_tensor_transpose_derivatives_extended_parameters — Function

tangenttensortransposederivativesextended_parameters(p::Vector{Float64})

Computes the derivatives of the tangent tensor's transpose with respect to the extended rotation parameters

Arguments

p::Vector{Float64} = rotation parameters

AeroBeams.rotation_parameters_WM — Function

rotationparametersWM(R::Matrix{<:Number})

Computes the Wiener-Milenkovic rotation parameters given a rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

AeroBeams.ypr_from_rotation_tensor — Function

yprfromrotation_tensor(R::Matrix{<:Number},tol::Float64=1e-6)

Computes the Euler angles from the sequence 3-2-1 given the rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

tol::Float64 = round-off tolerance

AeroBeams.quaternion_from_rotation_tensor — Function

quaternionfromrotation_tensor(R::Matrix{<:Number})

Computes the quaternion (Euler parameters) given a rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

AeroBeams.mode_tracking — Function

mode_tracking(controlParam::Vector{<:Number},freqs::Array{Vector{Float64}},damps::Array{Vector{Float64}},eigenvectors::Array{Matrix{ComplexF64}})

Applies mode tracking based on eigenvectors match

Arguments

controlParam::Vector{<:Number} = vector of control parameter

freqs::Array{Vector{Float64}} = frequencies vector

damps::Array{Vector{Float64}} = dampings vector

eigenvectors::Array{Matrix{ComplexF64}} = complex-valued eigenvectors

AeroBeams.get_FFT_and_PSD — Function

getFFTand_PSD(t::Vector{<:Number},y::Vector{<:Number},diffTol::Float64=1e3*eps())

Computes the FFT and PSD of signal y(t)

Arguments

t::Vector{<:Number} = time signal

y::Vector{<:Number} = quantity signal

diffTol::Float64 = tolerance for time signal being equally spaced

Visualizing the results

AeroBeams.plot_undeformed_assembly — Function

plotundeformedassembly(model::Model)

Plots the nodal coordinates of the assembly of beams

Arguments

model::Model

view::Tuple{<:Number,<:Number} = view angles

equalAspectRatio::Bool = flag to set equal aspect ratio plot

AeroBeams.plot_steady_deformation — Function

plotsteadydeformation(problem::Problem; plotBCs::Bool=true,view::Union{Nothing,Tuple{Int64,Int64}}=nothing,scale::Number=1,lw::Number=1,colorUndef=:black,colorDef=:blue,grid::Bool=true,legendPos=:best,tolPlane::Number=1e-8,plotAeroSurf::Bool=true,surfα::Float64=0.5,ΔuDef::Vector{<:Number}=zeros(3),save::Bool=false,savePath::String="/test/outputs/figures/fig.pdf")

Plots the initial and final deformed states for the model in the given problem

Arguments

problem::Problem = problem

Keyword arguments

plotBCs::Bool = flag to plot BCs

view::Union{Nothing,Tuple{Int64,Int64}} = view angles

scale::Number = displacements and rotations scale

lw::Number = linewidth

colorUndef = color for undeformed assembly

colorDef = color for deformed assembly

grid::Bool = flag for grid

legendPos = legend position

tolPlane::Number = displacement tolerance to plot as plane

plotAeroSurf = flag to plot aerodynamic surfaces

surfα::Float64 = transparency factor of aerodynamic surfaces

ΔuDef::Vector{<:Number} = displacement vector for first node of deformed assembly relative to the undeformed one

save::Bool = flag to save the figure

savePath::String = relative path on which to save the figure

AeroBeams.plot_steady_outputs — Function

plotsteadyoutputs(problem::Problem; outputs::Vector{String}=["u","p","F","M","V","Ω","α","cn","cm","ct","cl","cd"],beamGroups=1:length(problem.model.beams),lw::Number=1,colorScheme=:rainbow,legendPos=:best,save::Bool=false,saveFolder::String="/test/outputs/figures/",figureExtension::String=".pdf")

Plots outputs of a steady problem

Arguments

problem::Problem = problem

Keyword arguments

outputs::Vector{String} = list of outputs

beamGroups = list of beams in each group for which arclengths are concatenated

lw::Number = linewidth

colorScheme = color scheme

legendPos = legend position

save::Bool = flag to save figures

saveFolder::String = relative path of folder where to save figures

figureExtension::String = figure extension

AeroBeams.plot_mode_shapes — Function

plotmodeshapes(problem::Problem; plotBCs::Bool=true,view::Union{Nothing,Tuple{Int64,Int64}}=nothing,nModes::Union{Nothing,Int64}=nothing,scale::Number=1,frequencyLabel::String="frequency&damping",lw::Number=1,colorSteady=:black,modalColorScheme=:jet1,grid::Bool=true,legendPos=:best,tolPlane::Number=1e-8,plotAeroSurf::Bool=true,surfα::Float64=0.5,save::Bool=false,savePath::String="/test/outputs/figures/fig.pdf")

Plots the mode shapes of the model in the given problem

Arguments

problem::Problem = problem

Keyword arguments

plotBCs::Bool = flag to plot BCs

view::Union{Nothing,Tuple{Int64,Int64}} = view angles

nModes::Union{Nothing,Int64} = number of modes to plot

scale::Number = displacements and rotations scale

frequencyLabel::String = option for frequency label (only frequency or frequency and damping)

lw::Number = linewidth

colorSteady = color for steadily deformed assembly

modalColorScheme = color scheme for mode shapes

grid::Bool = flag for grid

legendPos = legend position

tolPlane::Number = displacement tolerance to plot as plane

plotAeroSurf = flag to plot aerodynamic surfaces

surfα::Float64 = transparency factor of aerodynamic surfaces

save::Bool = flag to save the figure

savePath::String = relative path on which to save the figure

AeroBeams.plot_dynamic_deformation — Function

plotdynamicdeformation(problem::Problem; refBasis::String="A", plotFrequency::Int64=1,plotUndeformed::Bool=false,plotBCs::Bool=true,plotDistLoads::Bool=true,view::Union{Nothing,Tuple{Int64,Int64}}=nothing,fps::Number=30,scale::Number=1,lw::Number=1,colorUndef=:black,colorDef=:blue,grid::Bool=true,legendPos=:best,tolPlane::Number=1e-8,plotAeroSurf::Bool=true,surfα::Float64=0.5,plotLimits::Union{Nothing,Vector{Tuple{T1,T2}}}=nothing,save::Bool=false,savePath::String="/test/outputs/figures/fig.gif",showScale::Bool=true,showTimeStamp::Bool=true,scalePos::Vector{<:Number}=[0.1;0.05;0.05],timeStampPos::Vector{<:Number}=[0.5;0.05;0.05],displayProgress::Bool=false) where {T1<:Number,T2<:Number}

Plots the animated deformation of the model in the given problem

Arguments

problem::Problem = problem

Keyword arguments

refBasis::String = reference basis for plot

plotFrequency::Int64 = frequency of time steps to plot

plotUndeformed::Bool = flag to plot undeformed assembly

plotBCs::Bool = flag to plot BCs

plotDistLoads::Bool = flag to plot distributed loads (includes gravitational and aerodynamic loads)

view::Union{Nothing,Tuple{Int64,Int64}} = view angles

fps::Number = frame rate for gif

scale::Number = displacements and rotations scale

frequencyLabel::String = option for frequency label (only frequency or frequency and damping)

lw::Number = linewidth

colorUndef = color for undeformed assembly

colorDef = color for deformed assembly

grid::Bool = flag for grid

legendPos = legend position

tolPlane::Number = displacement tolerance to plot as plane

plotAeroSurf = flag to plot aerodynamic surfaces

surfα::Float64 = transparency factor of aerodynamic surfaces

plotLimits::Union{Nothing,Vector{Tuple{T1,T2}}} = plot axis limits

save::Bool = flag to save the figure

savePath::String = relative path on which to save the figure

showScale::Bool = flag to show scale on plot

showTimeStamp::Bool = flag to show time stamp on plot

scalePos::Vector{<:Number} = position of scale on plot

timeStampPos::Vector{<:Number} = position of time stamp on plot

displayProgress::Bool = flag to display progress of gif creation

AeroBeams.plot_time_outputs — Function

plottimeoutputs(problem::Problem; nodes::Vector{Tuple{Int64,Int64}}=Vector{Tuple{Int64,Int64}}(),elements::Vector{Int64}=Vector{Int64}(),nodalOutputs::Vector{String}=["u","p","F","M"],elementalOutputs::Vector{String}=["u","p","F","M","V","Ω","α","cn","cm","ct","cl","cd"],lw::Number=1,colorScheme=:rainbow,showLegend::Bool=true,legendPos=:best,save::Bool=false,saveFolder::String="/test/outputs/figures/",figureExtension::String=".pdf")

Plots outputs of a dynamic problem

Arguments

problem::Problem = problem

Keyword arguments

nodes::Vector{Tuple{Int64,Int64}} = global IDs of nodes for which to plot

elements::Vector{Int64} = global IDs of elements for which to plot

nodalOutputs::Vector{String} = nodal outputs to plot

elementalOutputs::Vector{String} = elemental outputs to plot

lw::Number = linewidth

colorScheme = color scheme

showLegend::Bool = flag to show legend

legendPos = legend position

save::Bool = flag to save figures

saveFolder::String = relative path of folder where to save figures

figureExtension::String = figure extension