Public API

Creating an aerodynamic derivatives solver

AeroBeams.AD — Type

AD (Automatic Differentiation) DerivationMethod composite type

AeroBeams.FD — Type

FD (Finite Differences) DerivationMethod composite type

Fields

method::FiniteDifferenceMethod

Creating an aerodynamic solver

AeroBeams.QuasiSteady — Type

QuasiSteady AeroSolver composite type

AeroBeams.Indicial — Type

Indicial AeroSolver composite type

AeroBeams.BLi — Type

Incompressible modified Beddoes-Leishman AeroSolver composite type

AeroBeams.BLo — Type

Original Beddoes-Leishman AeroSolver composite type

AeroBeams.Inflow — Type

Inflow AeroSolver composite type

AeroBeams.TableLookup — Type

TableLookup FlapAeroSolver composite type

AeroBeams.ThinAirfoilTheory — Type

ThinAirfoilTheory FlapAeroSolver composite type

AeroBeams.IndicialGust — Type

IndicialGust GustAeroSolver composite type

Creating an airfoil

AeroBeams.create_Airfoil — Function

create_Airfoil(; kwargs...)

Airfoil constructor

Arguments

name::String = name of the airfoil

Re::Number = Reynolds number

Ma::Number = Mach number

U::Number = relative airspeed

b::Number = semichord

AeroBeams.create_flapped_Airfoil — Function

create_flapped_Airfoil(; kwargs...)

Airfoil constructor (with a trailing-edge flap)

Arguments

name::String = name of the airfoil

flapSiteID::Int64 = flap site ID

Re::Number = Reynolds number

Ma::Number = Mach number

U::Number = relative airspeed

b::Number = semichord

Creating an aerodynamic surface

AeroBeams.create_AeroSurface — Function

create_AeroSurface(; kwargs...)

Aerodynamic surface constructor

Keyword arguments

solver::AeroSolver = aerodynamic solver for pitch-plunge-induced loads

flapLoadsSolver::FlapAeroSolver = aerodynamic solver for flap-induced loads

gustLoadsSolver::GustAeroSolver = aerodynamic solver for gust-induced loads

derivationMethod::DerivationMethod = method for calculation of aerodynamic derivatives

airfoil::Airfoil = airfoil section

c::Union{<:Function,Number} = chord

Λ::Union{<:Function,Number} = sweep angle

normSparPos::Union{<:Function,Float64} = normalized position of the spar (beam reference line) on the chord

normFlapSpan::Union{Nothing,Vector{<:Number}} = normalized position of the trailing-edge flap along the span (beam arclength)

normFlapPos::Union{Nothing,Float64} = normalized position of the trailing-edge flap hinge on the chord

δIsTrimVariable::Bool = flag for trailing-edge deflection being a trim variable

δ::Union{Nothing,<:Function,Number} = trailing-edge deflection [rad]

flapSiteID::Union{Nothing,Int64} = trailing-edge flap site ID

updateAirfoilParameters::Bool = flag to update airfoil parameters with local airspeed

hasTipCorrection::Bool = flag to employ a tip correction on aerodynamic coefficients

tipLossFunction::String = respective tip loss function

tipLossDecayFactor::Float64 = respective tip loss factor

smallAngles::Bool = flag to employ small angles approximation on the calculation of the angle of attack

Creating an atmosphere

AeroBeams.standard_atmosphere — Function

standard_atmosphere(altitude::Number)

Atmosphere constructor based on the International Standard Atmosphere (https://en.wikipedia.org/wiki/InternationalStandardAtmosphere)

Arguments

altitude::Number

Creating boundary conditions

AeroBeams.create_BC — Function

create_BC(; kwargs...)

BC constructor

Keyword arguments

name::String = name of the BC

beam::Beam = beam at which the BC is applied

node::Int64 = node of the beam at which the BC is applied

types::Vector{String}= types of BCs applied to the node (generalized forces and displacements)

values = corresponding values of the applied BCs (constants or functions of time)

toBeTrimmed::Union{BitVector,Vector{Bool}} = TF on whether the BC is to be trimmed

Creating a beam

AeroBeams.create_Beam — Function

create_Beam(; kwargs...)

Beam constructor

Keyword Arguments

name::String = name of the beam

length::Number = (arc)length of the beam

rotationParametrization::String = type of rotation parametrization to define basis b

p0::Vector{<:Number} = rotation parameters from basis A to basis b

k::Vector{<:Number} = undeformed beam's curvatures per unit length

initialPosition::Vector{<:Number} = initial position of the beam's first node relative to the beam's origin (which may be another beam's node)

nElements::Int64 = number of elements for discretization

normalizedNodalPositions::Vector{Float64} = normalized nodal positions of beam elements

C::Vector{<:Matrix{<:Number}} = array of sectional stiffness matrices

I::Vector{<:Matrix{<:Number}} = array of sectional inertia matrices

connectedBeams::Union{Nothing,Vector{Beam}} = array of beams to which this beam is connected (a non-recursive property)

connectedNodesThis::Vector{Int64} = nodes of this beam which are connected to other beams' nodes (a non-recursive property)

connectedNodesOther::Vector{Int64} = respective nodes of the other beams (a non-recursive property)

pointInertias::Vector{PointInertia} = attached point inertias

hingedNodes::Vector{Int64} = nodes with a hinge

hingedNodesDoF::Union{Vector{Vector{Bool}},Vector{BitVector}} = respective hinged degrees-of-freedom

u0_of_x1::Union{Vector{<:Number},<:Function,Nothing} = initial displacement (resolved in the undeformed beam basis, b) of the beam as a function of its arclength coordinate (x1)

p0_of_x1::Union{Vector{<:Number},<:Function,Nothing} = initial rotation parameters (resolved in the undeformed beam basis, b) of the beam as a function of its arclength coordinate (x1)

udot0_of_x1::Union{Vector{<:Number},<:Function,Nothing} = initial displacement's rates (resolved in the undeformed beam basis, b) of the beam as a function of its arclength coordinate (x1)

pdot0_of_x1::Union{Vector{<:Number},<:Function,Nothing} = initial rotation parameters' rates (resolved in the undeformed beam basis, b) of the beam as a function of its arclength coordinate (x1)

f_A_of_x1t::Union{Nothing,<:Function} = distributed dead forces initially resolved in basis A, as a function of the beam arclength coordinate (x1) and time (t)

m_A_of_x1t::Union{Nothing,<:Function} = distributed dead moments initially resolved in basis A, as a function of the beam arclength coordinate (x1) and time (t)

f_b_of_x1t::Union{Nothing,<:Function} = distributed dead forces initially resolved in basis b, as a function of the beam arclength coordinate (x1) and time (t)

m_b_of_x1t::Union{Nothing,<:Function} = distributed dead moments initially resolved in basis b, as a function of the beam arclength coordinate (x1) and time (t)

ff_A_of_x1t::Union{Nothing,<:Function} = distributed follower forces initially resolved in basis A, as a function of the beam arclength coordinate (x1) and time (t)

mf_A_of_x1t::Union{Nothing,<:Function} = distributed moments forces initially resolved in basis A, as a function of the beam arclength coordinate (x1) and time (t)

ff_b_of_x1t::Union{Nothing,<:Function} = distributed follower forces initially resolved in basis b, as a function of the beam arclength coordinate (x1) and time (t)

mf_b_of_x1t::Union{Nothing,<:Function} = distributed follower moments initially resolved in basis b, as a function of the beam arclength coordinate (x1) and time (t)

aeroSurface::Union{Nothing,AeroSurface} = attached aerodynamic surface

springs::Vector{Spring} = array of attached springs

AeroBeams.update_beam! — Function

update_beam!(beam::Beam)

Validates and updates the beam construction

Arguments

beam::Beam

AeroBeams.add_point_inertias_to_beam! — Function

add_point_inertias_to_beam!(beam::Beam; inertias::Vector{PointInertia})

Adds point inertias to the beam

Arguments

beam::Beam

Keyword arguments

inertias::Vector{PointInertia}

AeroBeams.add_loads_to_beam! — Function

add_loads_to_beam!(beam::Beam; loadTypes::Vector{String},loadFuns::Vector{<:Function})

Adds loads to the beam

Arguments

beam::Beam

Keyword arguments

loadTypes::Vector{String}

loadFuns::Vector{<:Function}

AeroBeams.add_initial_displacements_and_velocities_to_beam! — Function

add_initial_displacements_and_velocities_to_beam!(beam::Beam; conditionTypes::Vector{String},conditionFuns::Vector{<:Function})

Adds initial generalized displacements and velocities to the beam

Arguments

beam::Beam

Keyword arguments

conditionTypes::Vector{String}

conditionFuns::Vector{<:Function}

AeroBeams.add_springs_to_beam! — Function

add_springs_to_beam!(; beam::Beam,springs::Vector{Spring})

Adds simply-attached springs to a beam

Keyword arguments

beam::Beam

springs::Vector{Spring}

AeroBeams.add_spring_to_beams! — Function

add_spring_to_beams!(; beams::Vector{Beam},spring::Spring)

Adds a doubly-attached spring to the beams

Keyword arguments

beams::Vector{Beam}

spring::Spring

Creating a gust

AeroBeams.create_SharpEdgedGust — Function

create_SharpEdgedGust(; kwargs...)

Creates a sharp-edged gust

Keyword arguments

initialTime::Number = time when the gust begins

duration::Number = duration of the gust

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

verticalVelocity::Number = peak "vertical" velocity of the gust (in lift direction)

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

AeroBeams.create_OneMinusCosineGust — Function

create_OneMinusCosineGust(; kwargs...)

Creates a one-minus-cosine gust

Keyword arguments

initialTime::Number = time when the gust begins

duration::Number = duration of the gust

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

verticalVelocity::Number = peak "vertical" velocity of the gust (in lift direction)

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

AeroBeams.create_Continuous1DGust — Function

create_Continuous1DGust(; kwargs...)

Creates a continuous 1D gust

Keyword arguments

spectrum::String = spectrum of the PSD: von Kármán ("vK") or Dryden ("Dryden")

generationMethod::String = method for generation of gust velocity ("sinusoids" or "whiteNoise")

initialTime::Number = time when the gust begins

duration::Number = duration of the gust

generationDuration::Number = duration of the gust considered for its generation

L::Number = turbulence length scale [m]

ωmin::Number = minimum frequency of the PSD [rad/s]

ωmax::Number = maximum frequency of the PSD [rad/s]

Uref::Number = reference airspeed

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

σ::Number = turbulence intensity (RMS) of "vertical" gust velocity component

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

seed::Int64 = seed for random numbers generation (reproducibility)

plotPSD::Bool = flag to plot the PSD

AeroBeams.create_DiscreteSpaceGust — Function

create_DiscreteSpaceGust(; kwargs...)

Creates a discrete space gust

Keyword arguments

type::String = type of gust

length::Number = length of the gust (in longitudinal direction)

width::Number = width of the gust (in spanwise direction)

convectiveVelocity::Number = convective velocity of the gust (in longitudinal direction)

verticalVelocity::Number = peak "vertical" velocity of the gust (in lift direction)

c0::Vector{<:Number} = position vector of the front of the gust, resolved in the inertial basis

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

AeroBeams.create_Continuous1DSpaceGust — Function

create_Continuous1DSpaceGust(; kwargs...)

Creates a continuous 1D space gust

Keyword arguments

spectrum::String = spectrum of the PSD: von Kármán ("vK") or Dryden ("Dryden")

length::Number = length of the gust (in longitudinal direction)

N::Int64 = number of nodes for length discretization

L::Number = turbulence length scale

σ::Number = turbulence intensity (RMS) of "vertical" gust velocity component

c0::Vector{<:Number} = position vector of the front of the gust, resolved in the inertial basis

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

seed::Int64 = seed for random numbers generation (reproducibility)

AeroBeams.create_Continuous2DSpaceGust — Function

create_Continuous2DSpaceGust(; kwargs...)

Creates a continuous 2D space gust

Keyword arguments

spectrum::String = spectrum of the PSD: von Kármán ("vK") or Dryden ("Dryden")

length::Number = length of the gust (in longitudinal direction)

width::Number = width of the gust (in lateral direction)

Nx::Int64 = number of nodes for length discretization

Ny::Int64 = number of nodes for width discretization

L::Number = turbulence length scale

σ::Number = turbulence intensity (RMS) of "vertical" gust velocity component

c0::Vector{<:Number} = position vector of the front of the gust, resolved in the inertial basis

p::Vector{<:Number} = Euler rotation parameters (3-2-1 sequence) from the inertial basis to the gust basis

seed::Int64 = seed for random numbers generation (reproducibility)

Creating a link

AeroBeams.create_TrimLoadsLink — Function

create_TrimLoadsLink(; masterBC::BC,slaveBCs::Vector{BC})

Trim loads link constructor (a link between trim loads so that they are equal)

Keyword arguments

masterBC::BC = master BC

slaveBCs::Vector{BC} = slave BCs

AeroBeams.create_FlapLink — Function

create_FlapLink(; kwargs...)

Flap link constructor (a link between flapped surfaces)

Keyword arguments

masterBeam::Beam = beam of the master surface

slaveBeams::Vector{Beam} = beams of the slave surfaces

δMultipliers::Vector{<:Number} = multiplication factors of flap deflection in slave surfaces relative to the master surface

Creating a model

AeroBeams.create_Model — Function

create_Model(; kwargs...)

Creates a model

Keyword arguments

name::String = name of the model

units::create_UnitsSystem = units system

beams::Vector{Beam} = beams in the assembly

initialPosition::Vector{<:Number} = initial position vector of the first node of the first beam, resolved in the inertial frame I

gravityVector::Vector{<:Number} = gravity vector

BCs::Vector{BC} = boundary condtions

p_A0::Vector{Float64} = initial rotation parameters that bring basis I to basis A

u_A::Union{Vector{<:Number},<:Function,Nothing} = displacement of basis A relative to basis I

v_A::Union{Vector{<:Number},<:Function,Nothing} = velocity of basis A relative to basis I

ω_A::Union{Vector{<:Number},<:Function,Nothing} = angular velocity of basis A relative to basis I

vdot_A::Union{Vector{<:Number},<:Function,Nothing} = acceleration of basis A relative to basis I

ωdot_A::Union{Vector{<:Number},<:Function,Nothing} = angular acceleration of basis A relative to basis I

altitude::Union{Nothing,Number} = altidude

atmosphere::Union{Nothing,Atmosphere} = atmosphere

gust::Union{Nothing,Gust} = gust

trimLoadsLinks::Vector{TrimLoadsLink} = links between trim loads

flapLinks::Vector{FlapLink} = links between flapped surfaces

rotationConstraints::Vector{RotationConstraint} = rotation constraints

AeroBeams.update_model! — Function

update_model!(model::Model)

Updates the model with its current settings

Arguments

model::Model

AeroBeams.set_motion_basis_A! — Function

set_motion_basis_A!(; kwargs...)

Sets the motion of basis A into the model

Keyword arguments

model::Model = model

u_A::Union{Vector{<:Number},<:Function,Nothing} = displacement of basis A relative to basis I

v_A::Union{Vector{<:Number},<:Function,Nothing} = velocity of basis A relative to basis I

ω_A::Union{Vector{<:Number},<:Function,Nothing} = angular velocity of basis A relative to basis I

vdot_A::Union{Vector{<:Number},<:Function,Nothing} = acceleration of basis A relative to basis I

ωdot_A::Union{Vector{<:Number},<:Function,Nothing} = angular acceleration of basis A relative to basis I

Creating a point inertia

AeroBeams.PointInertia — Type

PointInertia composite type

Fields

elementID::Int64 = local element ID to which the point inertia is attached

mass::Number = mass

η::Vector{Number} = position relative to element's midpoint's reference line

Ixx::Number = mass moment of inertia about the x1-axis

Iyy::Number = mass moment of inertia about the x2-axis

Izz::Number = mass moment of inertia about the x3-axis

Ixy::Number = mass product of inertia with respect to the x1-axis and x2-axis

Ixz::Number = mass product of inertia with respect to the x1-axis and x3-axis

Iyz::Number = mass product of inertia with respect to the x2-axis and x3-axis

inertiaMatrix::Union{Matrix{Number},Matrix{Nothing}} = mass moment of inertia matrix (tensor)

Creating and solving a problem

AeroBeams.InitialVelocitiesUpdateOptions — Type

InitialVelocitiesUpdateOptions composite type

Defines variables for the update of the initial velocities states

Fields

Δt::Number = time step

maxIter::Int64 = maximum number of iterations

relaxFactor::Float64 = relaxation factor

tol::Float64 = convergence tolerance

displayProgress::Bool = flag to display progress

AeroBeams.create_SteadyProblem — Function

create_SteadyProblem(; kwargs...)

Steady problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

x0::Vector{Float64} = initial states

AeroBeams.create_TrimProblem — Function

create_TrimProblem(; kwargs...)

Trim problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

getInertiaMatrix::Bool = flag to compute inertia matrix

x0::Vector{Float64} = initial states

AeroBeams.create_EigenProblem — Function

create_EigenProblem(; kwargs...)

Eigen problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

nModes::Int64=Inf64 = number of modes to be computed

frequencyFilterLimits::Vector{Float64} = limits of the frequency filter

normalizeModeShapes::Bool = flag to normalize mode shapes

x0::Vector{Float64} = initial states

jacobian::SparseMatrixCSC{Float64,Int64} = Jacobian matrix

inertia::SparseMatrixCSC{Float64,Int64} = inertia matrix

AeroBeams.create_DynamicProblem — Function

create_DynamicProblem(; kwargs...)

Dynamic problem constructor

Keyword arguments

model::Model = model

systemSolver::systemSolver = nonlinear system solver

getLinearSolution::Bool = flag to solve for linear structural solution

initialTime::Number = initial time

Δt::Union{Nothing,Number} = time step

finalTime::Union{Nothing,Number} = final time

timeVector::Union{Nothing,Vector{Float64}} = time vector

adaptableΔt::Bool = flag for adaptable time step

minΔt::Union{Nothing,Number} = minimum time step (when adaptable)

maxΔt::Union{Nothing,Number} = maximum time step (when adaptable)

δb::Float64 = discontinuities boundary convergence norm

skipInitialStatesUpdate::Bool = flag to skip update of initial states

initialVelocitiesUpdateOptions::InitialVelocitiesUpdateOptions = options for the initial velocities update

trackingTimeSteps::Bool = flag to track time step solutions

trackingFrequency::Int64 = frequency of time steps in which to track solution

displayProgress::Bool = flag to display progress

displayFrequency::Int64 = frequency of time steps in which to display progress

x0::Vector{Float64} = initial states

AeroBeams.solve! — Function

solve!(problem::Problem)

Solves a problem

Arguments

problem::Problem

AeroBeams.solve_eigen! — Function

solve_eigen!(problem::EigenProblem)

Solves an eigenproblem

Arguments

problem::EigenProblem

Creating a rotation constraint

AeroBeams.create_RotationConstraint — Function

create_RotationConstraint(; kwargs...)

Rotation constraint constructor

Keyword arguments

masterBeam::Beam = master beam

slaveBeam::Beam = slave beam

masterElementLocalID::Int64 = local ID of the master element

slaveElementLocalID::Int64 = local ID of the slave element

DOF::Int64 = constrained rotation degree-of-freedom

value::Number = value of the slave DOF relative to the master DOF

Creating a sample model

AeroBeams.tip_loss_factor_Pazy — Function

tip_loss_factor_Pazy(θ::Number,U::Number)

Computes the tip loss factor for the Pazy wing's tip correction function

Arguments

θ::Number = root pitch angle, in degrees

U::Number = airspeed

AeroBeams.geometrical_properties_Pazy — Function

geometrical_properties_Pazy()

Returns the fixed geometrical (and discretization) properties of the Pazy wing

AeroBeams.create_Pazy — Function

create_Pazy(; kwargs...)

Creates the Pazy wing

Returns the wing model and geometrical properties

Arguments

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

updateAirfoilParameters::Bool = flag to update airfoil parameters with airspeed

upright::Bool = flag to set the wing in the upright position

airfoil::Airfoil = airfoil section

θ::Real = pitch angle at the root [rad]

Λ::Real = sweep angle [rad]

withTipCorrection::Bool = flag for aerodynamic tip correction

GAy::Number = shear stiffness in the x2 direction

GAz::Number = shear stiffness in the x3 direction

altitude::Real = altitude

g::Real = acceleration of gravity

airspeed::Real = airspeed

tipMass::Real = mass of a point inertia added to the tip of the wing

ξtipMass::Vector{<:Real} = position vector of the tip mass relative to the tip of the spar, resolved in the local basis

tipMassInertia::Matrix{<:Real} = mass moment of inertia matrix of the tip mass

additionalBCs::Vector{BC} = additional BCs (beyond the clamp)

gust::Union{Nothing,Gust} = gust

AeroBeams.create_PazyFFWT — Function

create_PazyFFWT(; kwargs...)

Creates a version of the Pazy wing with flared folding wingtip (FFWT)

Arguments

p0::Vector{<:Number} = initial rotation parameters

airfoil::Airfoil = airfoil section

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

withTipCorrection::Bool = flag for aerodynamic tip correction

GAy::Number = shear stiffness in the x2 direction

GAz::Number = shear stiffness in the x3 direction

hingeNode::Int64 = hinge node

hingeAngle::Number = hinge (fold) angle

flareAngle::Number = flare angle

kSpring::Number = stiffness of the hinge

g::Number=0 = local acceleration of gravity

airspeed::Number = local airspeed

AeroBeams.create_SMW — Function

create_SMW(; kwargs...)

Creates the 16 meters wing (SMW) of the conventional HALE aircraft described by Patil, Hodges and Cesnik in: Nonlinear Aeroelasticity and Flight Dynamics of HALE (2001)

Returns the wing model and its span

Arguments

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

updateAirfoilParameters::Bool = flag to update airfoil parameters with airspeed

airfoil::Airfoil = airfoil section

θ::Real = pitch angle

k1::Real = twisting curvature

k2::Real = flapwise bending curvature

nElem::Int64 = number of elements for discretization

altitude::Real = altitude

airspeed::Real = airspeed

g::Real = acceleration of gravity

stiffnessFactor::Real = stiffness factor for beam structural properties

∞::Real = value of rigid structural properties

tipF3::Real = tip dead transverse force applied at the tip

AeroBeams.create_TDWing — Function

create_TDWing(; kwargs...)

Creates the wing described by Tang and Dowell in: Experimental and Theoretical Study on Aeroelastic Response of High-Aspect-Ratio Wings (2001)

Returns the wing model

Arguments

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

updateAirfoilParameters::Bool = flag to update airfoil parameters with airspeed

airfoil::Airfoil = airfoil section

θ::Real = pitch angle

nElem::Int64 = number of elements for discretization

airspeed::Real = airspeed

g::Real = acceleration of gravity

∞::Real = value of rigid structural properties

AeroBeams.create_Helios — Function

create_Helios(; kwargs...)

Creates a model based on the flying-wing aircraft described by Patil and Hodges in: Flight Dynamics of Highly Flexible Flying Wings (2006)

Keyword arguments

altitude::Number = altitude

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

g::Number = local acceleration of gravity

wingAirfoil::Airfoil = airfoil section of the wing

podAirfoil::Airfoil = airfoil section of the pods

beamPods::Bool = flag to include pods

stiffnessFactor::Number = stiffness factor for the wing structure

∞::Number = value for rigid structural properties

nElemStraightSemispan::Int64 = number of elements in the straight section of the semispan

nElemDihedralSemispan::Int64 = number of elements in the dihedral section of the semispan

nElemPod::Int64 = number of elements in the pods

payloadPounds::Number = payload, in pounds

airspeed::Number = local initial/trim airspeed

δIsTrimVariable::Bool = flag for flap deflection being a trim variable

thrustIsTrimVariable::Bool = flag for motors' thrust being a trim variable

δ::Union{Nothing,Number,<:Function} = flap deflection

thrust::Union{Number,<:Function} = motors' thrust

reducedChord::Bool = flag to employ a reduced (7 ft) chord

payloadOnWing::Bool = flag to set the payload on the wing's reference line

AeroBeams.create_conventional_HALE — Function

create_conventional_HALE(; kwargs...)

Creates a model based on the conventional HALE aircraft described by Patil, Hodges and Cesnik in: Nonlinear Aeroelasticity and Flight Dynamics of HALE (2001)

Keyword arguments

altitude::Number = altitude

aeroSolver::AeroSolver = aerodynamic solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

flapLoadsSolver::FlapAeroSolver = aerodynamic solver for flap loads

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

stabilizersAero::Bool = flag for stabilizers with aerodynamic surfaces

includeVS::Bool = flag to include a vertical stabilizer in the model

nElemWing::Int64 = number of elements of the full wing

nElemHorzStabilizer::Int64 = number of elements of the horizontal stabilizer

nElemTailBoom::Int64 = number of elements of the tail boom

nElemVertStabilizer::Int64 = number of elements of the vertical stabilizer

∞::Number=1e12 = value of rigid structural properties

stiffnessFactor::Number = stiffness factor for the wing structure

k1::Number = undeformed wing torsional curvature

k2::Number = undeformed wing flapwise bending curvature

airspeed::Number = local initial/trim airspeed

δElevIsTrimVariable::Bool = flag for elevator deflection being a trim variable

thrustIsTrimVariable::Bool = flag for motors' thrust being a trim variable

δElev::Union{Nothing,Number,<:Function} = elevator deflection

thrust::Union{Number,<:Function} = motors' thrust

g::Number = local acceleration of gravity

wingCd0::Number = parisite drag coefficient for the wing

wingcnδ::Number = cn vs δ slope for the wing

wingcmδ::Number = cm vs δ slope for the wing

wingcdδ::Number = cd vs δ slope for the wing

stabsCd0::Number = parisite drag coefficient for the stabilizers

stabscnδ::Number = cn vs δ slope for the stabilizers

stabscmδ::Number = cm vs δ slope for the stabilizers

stabscdδ::Number = cd vs δ slope for the stabilizers

AeroBeams.create_BWB — Function

create_BWB(; kwargs...)

Creates a model based on the blended-wing-body described by Weihua Su's PhD thesis

Keyword arguments

altitude::Number = altitude

aeroSolver::AeroSolver = aerodynamic solver

gustLoadsSolver::GustAeroSolver = indicial gust loads solver

derivationMethod::DerivationMethod = method for aerodynamic derivatives

∞::Number=1e12 = value of rigid structural properties

stiffnessFactor::Number = stiffness factor for the wing structure

airspeed::Number = local initial/trim airspeed

δElevIsTrimVariable::Bool = flag for elevator deflection being a trim variable

thrustIsTrimVariable::Bool = flag for motors' thrust being a trim variable

δElev::Union{Nothing,Number,<:Function} = elevator deflection

thrust::Union{Number,<:Function} = motors' thrust

g::Number = local acceleration of gravity

updateAirfoilParameters::Bool = flag to update airfoil parameters with airspeed

hasTipCorrection::Bool = flag to employ aerodynamic tip correction

tipLossDecayFactor::Number = tip loss decay factor

Creating a spring

AeroBeams.create_Spring — Function

create_Spring(; kwargs...)

Creates a spring

Keyword arguments

basis::String = basis on which stiffnesses are defined

elementsIDs::Vector{Int64} = local IDs of the element(s)' node(s) to which the spring is attached

nodesSides::Vector{Int64} = sides (1 or 2) of the node(s) to which the spring is attached

ku::Vector{<:Number} = translational stiffness vector

kp::Vector{<:Number} = rotational stiffness vector

kTranslational::Number = translational stiffness (same in all directions)

kTwist::Number = twist stiffness (rotation about x1 direction)

kIPBending::Number = in-plane bending stiffness (rotation about x3 direction)

kOOPBending::Number = out-of-plane bending stiffness (rotation about x2 direction)

Creating a system solver

AeroBeams.create_NewtonRaphson — Function

create_NewtonRaphson(; kwargs...)

Newton-Raphson nonlinear system solver constructor

Keyword arguments

absoluteTolerance::Float64 = absolute convergence tolerance

relativeTolerance::Float64 = relative convergence tolerance

maximumIterations::Int64 = maximum number of iterations

desiredIterations::Int64 = desired number of iterations

maximumAbsoluteError::Number = maximum absolute error for divergence detection

maximumRelativeError::Number = maximum relative error for divergence detection

initialLoadFactor::Number = initial load factor

minimumLoadFactor::Float64 = minimum load factor

maximumLoadFactorStep::Float64 = maximum load factor step

minimumLoadFactorStep::Float64 = minimum load factor step

ρ::Float64 = relaxation factor for trim variables

trackingLoadSteps::Bool = flag to track partial load steps solutions

displayStatus::Bool = flag to display status

minConvRateAeroJacUpdate::Number = minimum convergence rate to skip computation of aerodynamic Jacobians

minConvRateJacUpdate::Number = minimum convergence rate to skip computation of structural Jacobians

alwaysUpdateJacobian::Bool = flag to update Jacobians on every iteration

Creating a units system

AeroBeams.create_UnitsSystem — Function

create_UnitsSystem(; kwargs...)

Creates a system composed of length, force, angle, frequency and mass units (this is only for plotting purposes and does not influence calculations)

Keyword arguments

length::String = length unit

force::String = force unit

angle::String = angle unit

frequency::String = frequency unit

mass::String = mass unit

Utilities

AeroBeams.round_off! — Function

round_off!(x)

Rounds the array or number to input tolerance (defaults to machine epsilon)

Arguments

x = array / number

tol = tolerance

AeroBeams.rms — Function

rms(x)

Computes the root mean square of an array

AeroBeams.tilde — Function

tilde(v::Vector{<:Number})

Computes the skew-symmetric matrix associated with a vector

Arguments

v::Vector{<:Number} = three-element vector

AeroBeams.isotropic_stiffness_matrix — Function

isotropic_stiffness_matrix(; kwargs...)

Creates a 6x6 sectional stiffness matrix

Arguments

∞::Number = value for rigid properties

EA::Number = axial stiffness

GAy::Number = shear stiffness in the x2 direction

GAz::Number = shear stiffness in the x3 direction

GJ::Number = torsional stiffness

EIy::Number = bending stiffness about the x2 direction

EIz::Number = bending stiffness about the x3 direction

AeroBeams.inertia_matrix — Function

inertia_matrix(; kwargs...)

Creates a 6x6 sectional inertia matrix

Arguments

ρA::Number = mass per unit length

ρIy::Number = mass moment of inertia per unit length about the x2-axis

ρIz::Number = mass moment of inertia per unit length about the x3-axis

ρIs::Number = mass moment of inertia per unit length about the x1-axis

e2::Number = offset of center of gravity in the x2 direction with respect to the local reference line

e3::Number = offset of center of gravity in the x3 direction with respect to the local reference line

AeroBeams.rotation_tensor_E321 — Function

rotation_tensor_E321(p::Vector{<:Number})

Computes the rotation tensor according to Euler parameters sequence 3-2-1

Arguments

p::Vector{<:Number} = rotation parameters

AeroBeams.rotation_tensor_E313 — Function

rotation_tensor_E313(p::Vector{<:Number})

Computes the rotation tensor according to Euler parameters sequence 3-1-3

Arguments

p::Vector{<:Number} = rotation parameters

AeroBeams.rotation_tensor_WM — Function

rotation_tensor_WM(p::Vector{<:Number})

Computes the rotation tensor according to Wiener-Milenkovic parameters

Arguments

p::Vector{<:Number} = rotation parameters

AeroBeams.rotation_parameters_WM — Function

rotation_parameters_WM(R::Matrix{<:Number})

Computes the Wiener-Milenkovic rotation parameters given a rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

AeroBeams.ypr_from_rotation_tensor — Function

ypr_from_rotation_tensor(R::Matrix{<:Number},tol::Float64=1e-6)

Computes the Euler angles from the sequence 3-2-1 given the rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

tol::Float64 = round-off tolerance

AeroBeams.quaternion_from_rotation_tensor — Function

quaternion_from_rotation_tensor(R::Matrix{<:Number})

Computes the quaternion (Euler parameters) given a rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

AeroBeams.mode_tracking — Function

mode_tracking(controlParam::Vector{<:Number},freqs::Array{Vector{Float64}},damps::Array{Vector{Float64}},eigenvectors::Array{Matrix{ComplexF64}})

Applies mode tracking based on eigenvectors match

Arguments

controlParam::Vector{<:Number} = vector of control parameter

freqs::Array{Vector{Float64}} = frequencies vector

damps::Array{Vector{Float64}} = dampings vector

eigenvectors::Array{Matrix{ComplexF64}} = complex-valued eigenvectors

AeroBeams.get_FFT_and_PSD — Function

get_FFT_and_PSD(t::Vector{<:Number},y::Vector{<:Number}; tol::Float64=1e3*eps())

Computes the FFT and PSD of signal y(t)

Arguments

t::Vector{<:Number} = time signal

y::Vector{<:Number} = quantity signal

Keyword arguments

tol::Float64 = tolerance for time signal being equally spaced

Visualizing the results

AeroBeams.plot_undeformed_assembly — Function

plot_undeformed_assembly(model::Model,view::Tuple{Int64,Int64}=(45,45),equalAspectRatio::Bool=true)

Plots the nodal coordinates of the assembly of beams

Arguments

model::Model

view::Tuple{<:Number,<:Number} = view angles

equalAspectRatio::Bool = flag to set equal aspect ratio plot

AeroBeams.plot_steady_deformation — Function

plot_steady_deformation(problem::Problem; kwargs...)

Plots the initial and final deformed states for the model in the given problem

Arguments

problem::Problem = problem

Keyword arguments

plotBCs::Bool = flag to plot BCs

view::Union{Nothing,Tuple{Int64,Int64}} = view angles

scale::Number = displacements and rotations scale

lw::Number = linewidth

colorUndef = color for undeformed assembly

colorDef = color for deformed assembly

grid::Bool = flag for grid

legendPos = legend position

tolPlane::Number = displacement tolerance to plot as plane

plotAeroSurf = flag to plot aerodynamic surfaces

surfα::Float64 = transparency factor of aerodynamic surfaces

ΔuDef::Vector{<:Number} = displacement vector for first node of deformed assembly relative to the undeformed one

plotLimits::Union{Nothing,Vector{Tuple{T1,T2}}} = plot axis limits

showScale::Bool = flag to show scale on plot

scalePos::Vector{<:Number} = position of scale on plot

save::Bool = flag to save the figure

savePath::String = relative path on which to save the figure

AeroBeams.plot_steady_outputs — Function

plot_steady_outputs(problem::Problem; kwargs...)

Plots outputs of a steady problem

Arguments

problem::Problem = problem

Keyword arguments

outputs::Vector{String} = list of outputs

beamGroups = list of beams in each group for which arclengths are concatenated

lw::Number = linewidth

colorScheme = color scheme

legendPos = legend position

save::Bool = flag to save figures

saveFolder::String = relative path of folder where to save figures

figureExtension::String = figure extension

AeroBeams.plot_mode_shapes — Function

plot_mode_shapes(problem::Problem; kwargs...)

Plots the mode shapes of the model in the given problem

Arguments

problem::Problem = problem

Keyword arguments

plotBCs::Bool = flag to plot BCs

view::Union{Nothing,Tuple{Int64,Int64}} = view angles

nModes::Union{Nothing,Int64} = number of modes to plot

scale::Number = displacements and rotations scale

frequencyLabel::String = option for frequency label (only frequency or frequency and damping)

lw::Number = linewidth

colorSteady = color for steadily deformed assembly

modalColorScheme = color scheme for mode shapes

grid::Bool = flag for grid

legendPos = legend position

tolPlane::Number = displacement tolerance to plot as plane

plotAeroSurf = flag to plot aerodynamic surfaces

surfα::Float64 = transparency factor of aerodynamic surfaces

save::Bool = flag to save the figure

savePath::String = relative path on which to save the figure

AeroBeams.plot_dynamic_deformation — Function

plot_dynamic_deformation(problem::Problem; kwargs...)

Plots the animated deformation of the model in the given problem

Arguments

problem::Problem = problem

Keyword arguments

refBasis::String = reference basis for plot

plotFrequency::Int64 = frequency of time steps to plot

plotUndeformed::Bool = flag to plot undeformed assembly

plotBCs::Bool = flag to plot BCs

plotDistLoads::Bool = flag to plot distributed loads (includes gravitational and aerodynamic loads)

view::Union{Nothing,Tuple{Int64,Int64}} = view angles

fps::Number = frame rate for gif

scale::Number = displacements and rotations scale

frequencyLabel::String = option for frequency label (only frequency or frequency and damping)

lw::Number = linewidth

colorUndef = color for undeformed assembly

colorDef = color for deformed assembly

grid::Bool = flag for grid

legendPos = legend position

tolPlane::Number = displacement tolerance to plot as plane

plotAeroSurf = flag to plot aerodynamic surfaces

surfα::Float64 = transparency factor of aerodynamic surfaces

plotLimits::Union{Nothing,Vector{Tuple{T1,T2}}} = plot axis limits

save::Bool = flag to save the figure

savePath::String = relative path on which to save the figure

showScale::Bool = flag to show scale on plot

showTimeStamp::Bool = flag to show time stamp on plot

scalePos::Vector{<:Number} = position of scale on plot

timeStampPos::Vector{<:Number} = position of time stamp on plot

displayProgress::Bool = flag to display progress of gif creation

AeroBeams.plot_time_outputs — Function

plot_time_outputs(problem::Problem; kwargs...)

Plots outputs of a dynamic problem

Arguments

problem::Problem = problem

Keyword arguments

nodes::Vector{Tuple{Int64,Int64}} = global IDs of nodes for which to plot

elements::Vector{Int64} = global IDs of elements for which to plot

nodalOutputs::Vector{String} = nodal outputs to plot

elementalOutputs::Vector{String} = elemental outputs to plot

lw::Number = linewidth

colorScheme = color scheme

showLegend::Bool = flag to show legend

legendPos = legend position

save::Bool = flag to save figures

saveFolder::String = relative path of folder where to save figures

figureExtension::String = figure extension