AeroBeams.AeroBeams

AeroBeams.I3

AeroBeams.I6

AeroBeams.a1

AeroBeams.BC

AeroBeams.Beam

AeroBeams.ComplementaryElementalStates

AeroBeams.ComplementaryElementalStatesRates

AeroBeams.DynamicProblem

AeroBeams.EigenProblem

AeroBeams.Element

AeroBeams.ElementalStates

AeroBeams.ElementalStatesRates

AeroBeams.InitialVelocitiesUpdateOptions

AeroBeams.ModeShape

AeroBeams.Model

AeroBeams.NewtonRaphson

AeroBeams.NodalStates

AeroBeams.PointInertia

AeroBeams.Problem

AeroBeams.SpecialNode

AeroBeams.SteadyProblem

AeroBeams.SystemSolver

AeroBeams.TrimLoadsLink

AeroBeams.TrimProblem

AeroBeams.UnitSystem

AeroBeams.add_initial_displacements_and_velocities_to_beam!

AeroBeams.add_loads_to_beam!

AeroBeams.add_point_inertia_to_element!

AeroBeams.add_point_inertias_to_beam!

AeroBeams.assemble_model!

AeroBeams.assemble_system_arrays!

AeroBeams.copy_initial_states!

AeroBeams.create_beam_elements!

AeroBeams.curvature_quantities

AeroBeams.distributed_external_loads

AeroBeams.distributed_external_loads_Jacobians

AeroBeams.distributed_loads_Jacobians!

AeroBeams.divide_inplace

AeroBeams.element_accelerations_basis_b!

AeroBeams.element_arrays!

AeroBeams.element_distributed_loads!

AeroBeams.element_inertia!

AeroBeams.element_jacobian!

AeroBeams.element_modal_states

AeroBeams.element_momenta!

AeroBeams.element_momenta_rates!

AeroBeams.element_nodal_states!

AeroBeams.element_residual!

AeroBeams.element_rotation_variables!

AeroBeams.element_states!

AeroBeams.element_states_rates!

AeroBeams.element_strains!

AeroBeams.element_velocities_basis_b!

AeroBeams.force_scaling

AeroBeams.get_element_distributed_loads

AeroBeams.get_equivalent_states_rates!

AeroBeams.get_hinged_nodes_matrices

AeroBeams.get_initial_displacements_derivatives!

AeroBeams.get_mode_shapes!

AeroBeams.get_rotation_tensors!

AeroBeams.get_special_nodes!

AeroBeams.get_system_indices!

AeroBeams.get_velocity_dofs_to_update!

AeroBeams.gravitational_loads

AeroBeams.gravitational_loads_Jacobians

AeroBeams.initialize_basis_A_rotation!

AeroBeams.initialize_system_arrays!

AeroBeams.initialize_time_variables!

AeroBeams.interpolate_distributed_loads

AeroBeams.mul3

AeroBeams.plot_undeformed_assembly

AeroBeams.position_vector_from_curvature

AeroBeams.precompute_distributed_loads!

AeroBeams.quaternion_from_rotation_tensor

AeroBeams.rotation_angle

AeroBeams.rotation_parameter_scaling

AeroBeams.rotation_parameters_WM

AeroBeams.rotation_tensor_E313

AeroBeams.rotation_tensor_E321

AeroBeams.rotation_tensor_WM

AeroBeams.rotation_tensor_derivatives_extended_parameters

AeroBeams.rotation_tensor_derivatives_extended_parameters

AeroBeams.rotation_tensor_derivatives_scaled_parameters

AeroBeams.rotation_tensor_derivatives_time_extended_parameters

AeroBeams.rotation_tensor_from_curvature

AeroBeams.rotation_tensor_time_derivative

AeroBeams.rotation_tensor_time_derivative

AeroBeams.round_off!

AeroBeams.save_load_factor_data!

AeroBeams.save_time_step_data!

AeroBeams.scaling_derivatives_extended_parameters

AeroBeams.set_BCs!

AeroBeams.set_initial_states!

AeroBeams.set_nodal_coordinates!

AeroBeams.solve!

AeroBeams.solve_NewtonRaphson!

AeroBeams.solve_dynamic!

AeroBeams.solve_eigen!

AeroBeams.solve_initial_dynamic!

AeroBeams.solve_linear_system!

AeroBeams.solve_steady!

AeroBeams.solve_time_step!

AeroBeams.special_node_arrays!

AeroBeams.special_node_follower_loads_jacobians!

AeroBeams.special_node_jacobian!

AeroBeams.special_node_residual!

AeroBeams.special_node_states!

AeroBeams.tangent_operator_transpose_WM

AeroBeams.tangent_operator_transpose_WM

AeroBeams.tangent_operator_transpose_inverse_WM

AeroBeams.tangent_operator_transpose_inverse_WM

AeroBeams.tangent_tensor_functions_derivatives_extended_parameters

AeroBeams.tangent_tensor_transpose_derivatives_extended_parameters

AeroBeams.tilde

AeroBeams.time_march!

AeroBeams.update_BC_data!

AeroBeams.update_basis_A_orientation!

AeroBeams.update_beam!

AeroBeams.update_element_distributed_loads!

AeroBeams.update_initial_conditions!

AeroBeams.update_initial_velocities!

AeroBeams.update_loads_trim_links_global_ids!

AeroBeams.update_model!

AeroBeams.update_special_node_jacobian!

AeroBeams.update_states!

AeroBeams.update_time_variables!

AeroBeams.validate_and_update_motion_basis_A!

AeroBeams.validate_beam!

AeroBeams.validate_connected_beams

AeroBeams.validate_distributed_loads!

AeroBeams.validate_hinged_nodes!

AeroBeams.validate_initial_conditions!

AeroBeams.validate_model!

AeroBeams.validate_normalized_nodal_positions!

AeroBeams.validate_rotation_parametrization

AeroBeams.validate_sectional_matrices

AeroBeams.validate_unit_system

AeroBeams.AeroBeams — Module

Module for the AeroBeams package

AeroBeams.I3 — Constant

I3 = Matrix(1.0*LinearAlgebra.I,3,3)

3x3 identity matrix

AeroBeams.I6 — Constant

I6 = Matrix(1.0*LinearAlgebra.I,6,6)

6x6 identity matrix

AeroBeams.a1 — Constant

const a1 = [1.0; 0.0; 0.0]

First vector of basis A, resolved in that basis

AeroBeams.BC — Type

@with_kw mutable struct BC

Boundary conditions composite type

Fields

name::String = name of the BC

beam::Beam = beam at which the BC is applied

node::Int64 = node of the beam at which the BC is applied

types::Vector{String} = types of BCs applied to the node (generalized forces and displacements)

values::Vector{Union{<:Number,<:Function}} = corresponding values of the applied BCs (constants or functions of time)

AeroBeams.Beam — Type

@with_kw mutable struct Beam

Beam composite type

Fields

name::String = name of the beam

connectedNodesThis::Int64 = which node is connected to another beam

connectedBeams::Union{Nothing, Beam} = to which beam the current beam is connected

connectedNodesOther::Int64 = to which node of the other beam it is connected

length::Float64 = total arclength of the beam

rotationParametrization::String = rotation parametrization for the undeformed geometry definition

p0::Vector{Float64} = corresponding rotation parameters

k::Vector{Float64} = initial curvatures of the beam in the undeformed geometry (torsional, flapwise bending, in-plane bending)

nElements::Int64 = number of elements

normalizedNodalPositions::Vector{Float64} = nodal positions normalized by the length

elementRange::Vector{Int64} = current beam's range of elements in the model

constitutiveRelation::{String} = constitutive relation of the material

C::Matrix{Float64} = sectional stiffness matrix

I::Matrix{Float64} = sectional inertia matrix

Notes

Some fields have meaningful default values, others need inputs

AeroBeams.ComplementaryElementalStates — Type

mutable struct ComplementaryElementalStates

ComplementaryElementalStates composite type

Fields

AeroBeams.ComplementaryElementalStatesRates — Type

mutable struct ComplementaryElementalStatesRates

ComplementaryElementalStatesRates composite type

Fields

AeroBeams.DynamicProblem — Type

@with_kw mutable struct DynamicProblem <: Problem

DynamicProblem composite type

Defines the problem of dynamic type

Fields

AeroBeams.EigenProblem — Type

@with_kw mutable struct EigenProblem <: Problem

EigenProblem composite type

Defines the problem of eigen type

Fields

AeroBeams.Element — Type

mutable struct Element <: BeamElement

Element composite type

Fields

parent::Beam

localID::Int64

Notes

Finite elements belong to a Beam

AeroBeams.ElementalStates — Type

mutable struct ElementalStates

ElementalStates composite type

Fields

AeroBeams.ElementalStatesRates — Type

mutable struct ElementalStatesRates

ElementalStatesRates composite type

Fields

AeroBeams.InitialVelocitiesUpdateOptions — Type

@with_kw mutable struct InitialVelocitiesUpdateOptions

InitialVelocitiesUpdateOptions composite type

Defines variables for the update of the initial velocities states

Fields

Δt::Number

maxIter::Int64

relaxFactor::Float64

tol::Float64

AeroBeams.ModeShape — Type

mutable struct ModeShape{T<:Union{Float64,ComplexF64}}

ModeShape composite type

Fields

mode::Int64

frequency::Float64

damping::Float64

elementalStates::Vector{ElementalStates{T}}

complementaryElementalStates::Vector{ComplementaryElementalStates{T}}

nodalStates::Vector{NodalStates{T}}

AeroBeams.Model — Type

@with_kw mutable struct Model

Model composite type

Fields

name::String = name of the model

units::UnitSystem = unit system (length, force, angle and frequency)

beams::Vector{Beam} = beams that compose the model's assembly

initialPosition::Vector{<:Number} = initial position of the first node of the first beam of the model, defined in basis A

gravityVector::Vector{<:Number} = gravity vector, defined in the I (inertial) frame

BCs::Dict{Int64,Int64} = boundary conditions applied to the model

nElementsTotal::Int64 = total number of elements of the beam's assembly

nNodesTotal::Int64 = total number of nodes of the beam's assembly

elementNodes::Vector{Vector{Int64}} = a nElementsTotal x 2 matrix in which the row represents the element, and the columns are the corresponding element's nodes

p_A0::Vector{Float64} = initial Euler 3-2-1 rotation parameters that bring basis I to basis A

R_A::Matrix{Float64}= 1.0I(3) = initial rotation tensor that brings basis I to basis A, resolved in basis A

nTrimVariables::Int64 = number of trim variables

trimLoadsLinks::Vector{TrimLoadsLink} = trim links among loads

Notes

The default is an empty model

AeroBeams.NewtonRaphson — Type

@with_kw mutable struct NewtonRaphson <: SystemSolver

Newton-Raphson system solver composite type

Fields

AeroBeams.NodalStates — Type

mutable struct NodalStates

NodalStates composite type

Fields

AeroBeams.PointInertia — Type

@with_kw mutable struct PointInertia

PointInertia composite type

Fields

elementID::Int64

mass::Number

η::Vector{Number} = position relative to element's midpoint

Ixx::Number

Iyy::Number

Izz::Number

Ixy::Number

Ixz::Number

Iyz::Number

inertiaMatrix::Union{Matrix{Number},Matrix{Nothing}}

AeroBeams.Problem — Type

abstract type Problem

Defines a general Problem type

AeroBeams.SpecialNode — Type

@with_kw mutable struct SpecialNode

Special node composite type

Fields

AeroBeams.SteadyProblem — Type

@with_kw mutable struct SteadyProblem <: Problem

SteadyProblem composite type

Defines the problem of steady type

Fields

AeroBeams.SystemSolver — Type

abstract type SystemSolver

Defines a general problem for the system of equations

AeroBeams.TrimLoadsLink — Type

@with_kw mutable struct TrimLoadsLink

TrimLoadsLink composite type

Fields

masterBC::BC

slaveBCs::Vector{BC}

masterBeam::Beam

masterNodeLocalID::Int64

slaveBeams::Vector{Beam}

slaveNodesLocalIDs::Vector{Int64}

masterNodeGlobalID::Int64

slaveNodesGlobalIDs::Vector{Int64}

AeroBeams.TrimProblem — Type

@with_kw mutable struct TrimProblem <: Problem

TrimProblem composite type

Defines the problem of trim type

Fields

AeroBeams.UnitSystem — Type

@with_kw mutable struct UnitSystem

A composite type with fields for length, force, angle and frequency units (this is only for plotting purposes and does not influence calculations)

Fields:

length::String

force::String

angle::String

frequency::String

AeroBeams.add_initial_displacements_and_velocities_to_beam! — Method

addinitialdisplacementsandvelocitiestobeam!(beam::Beam,conditionTypes::Vector{String},conditionFuns::Vector{<:Function})

Adds initial generalized displacements and velocities to the beam

Arguments

beam::Beam

conditionTypes::Vector{String}

conditionFuns::Vector{<:Function}

AeroBeams.add_loads_to_beam! — Method

addloadsto_beam!(;beam::Beam,loadTypes::Vector{String},loadFuns::Vector{<:Function})

Adds loads to the beam

Arguments

beam::Beam

loadTypes::Vector{String}

loadFuns::Vector{<:Function}

AeroBeams.add_point_inertia_to_element! — Method

addpointinertiatoelement!(element::Element,pointInertia::PointInertia)

Adds the point inertia's matrix to the element's sectional inertia matrix

Arguments

element::Element

pointInertia::PointInertia

AeroBeams.add_point_inertias_to_beam! — Method

addpointinertiastobeam!(;beam::Beam,inertias::Vector{PointInertia})

Adds point inertias to the beam

Arguments

beam::Beam

inertias::Vector{PointInertia}

AeroBeams.assemble_model! — Method

assemble_model!(model::Model,beams::Vector{Beam})

Loads the assembly of beams into the model

Arguments

model::Model = model at hand

beams::Vector{Beam} = beams to load into the model

Notes

AeroBeams.assemble_system_arrays! — Method

assemblesystemarrays!(problem::Problem)

Assembles the residual vector, Jacobian and inertia matrices of the system of equations

Arguments

problem::Problem

AeroBeams.copy_initial_states! — Method

copyinitialstates!(problem::Problem,problemCopy::Problem)

Copies the initial states to the original problem

Arguments

problem::Problem

problemCopy::Problem

AeroBeams.create_beam_elements! — Method

createbeamelements!(beam::Beam)

Initializes the element and node ranges, and creates beam elements

Arguments

beam::Beam

AeroBeams.curvature_quantities — Method

curvature_quantities(k::Vector{Float64})

Gets some useful functions of the curvature vector

Arguments

k::Vector{Float64} = curvature vector

AeroBeams.distributed_external_loads — Method

distributedexternalloads(problem::Problem,element::Element,σ::Float64)

Computes the nodal resultants from the externally applied distributed loads on the current element

Arguments

model::Model

element::Element

σ::Float64

AeroBeams.distributed_external_loads_Jacobians — Method

distributedexternalloads_Jacobians(element::Element)

Computes the contributions of the externally applied distributed loads to the Jacobian matrix

Arguments

element::Element

AeroBeams.distributed_loads_Jacobians! — Method

distributedloadsJacobians!(element::Element)

Computes the contributions of the distributed loads to the Jacobian matrix

Arguments

element::Element

AeroBeams.divide_inplace — Method

divide_inplace(divisor::Number, vars...)

Divides the input variables in-place

Arguments

divisor::Number = divisor

vars... = variables to be divided

AeroBeams.element_accelerations_basis_b! — Function

elementaccelerationsbasis_b!(model::Model,element::Element,σ::Float64=1.0,timeNow::Float64=0.0)

Gets the generalized accelerations of basis b at the element's midpoint, resolved in basis A

Arguments

model::Model

element::Element

σ::Float64 = load factor

timeNow::Float64 = current time

AeroBeams.element_arrays! — Method

element_arrays!(problem::Problem,model::Model,element::Element)

Gets the elemental contributions to the system's arrays (residual, Jacobian, inertia)

Arguments

problem::Problem

model::Model

element::Element

AeroBeams.element_distributed_loads! — Method

elementdistributedloads!(problem::Problem,model::Model,element::Element)

Gets the nodal resultants from distributed loads on the current element, resolved in basis A

Arguments

problem::Problem

model::Model

element::Element

AeroBeams.element_inertia! — Method

element_inertia!(problem::Problem,model::Model,element::Element)

Computes the contributions from the current element to the inertia matrix

Arguments

problem::Problem

model::Model

element::Element

AeroBeams.element_jacobian! — Method

element_jacobian!(problem::Problem,model::Model,element::Element)

Computes the contributions from the current element to the Jacobian matrix

Arguments

problem::Problem

model::Model

element::Element

AeroBeams.element_modal_states — Method

elementmodalstates(element::Element,eigenvector::Vector{T},forceScaling::Float64)

Gets the modal states (generalized displacements, forces, strains, velocities and momenta) of the element at midpoint and at the nodes

Arguments

element::Element

eigenvector::Vector{T}

forceScaling::Float64

AeroBeams.element_momenta! — Method

element_momenta!(element::Element)

Computes the momenta for the current element, resolved in basis B

Arguments

element::Element

AeroBeams.element_momenta_rates! — Method

elementmomentarates!(element::Element)

Computes the momenta rates for the current element, resolved in basis B

Arguments

element::Element

AeroBeams.element_nodal_states! — Method

elementnodalstates!(element::Element)

Updates the nodal states of the element

Arguments

element::Element

AeroBeams.element_residual! — Method

element_residual!(problem::Problem,model::Model,element::Element)

Computes the contributions from the current element to the residual array

Arguments

problem::Problem

model::Model

element::Element

AeroBeams.element_rotation_variables! — Method

elementrotationvariables!(problem::Problem,element::Element)

Gets the rotation variables for the current element

Arguments

problem::Problem

element::Element

AeroBeams.element_states! — Method

element_states!(problem::Problem,model::Model,element::Element)

Gets the states (generalized displacements, forces and velocities) of the element

Arguments

problem::Problem

model::Model

element::Element

AeroBeams.element_states_rates! — Method

elementstatesrates!(problem::Problem,element::Element)

Gets the states' rates of the current element

Arguments

problem::Problem

element::Element

AeroBeams.element_strains! — Method

element_strains!(element::Element)

Computes the strains for the current element, resolved in basis B

Arguments

element::Element

AeroBeams.element_velocities_basis_b! — Function

elementvelocitiesbasis_b!(model::Model,element::Element,σ::Float64=1.0,timeNow::Float64=0.0)

Gets the generalized velocities of basis b at the element's midpoint, resolved in basis A

Arguments

model::Model

element::Element

σ::Float64 = load factor

timeNow::Float64 = current time

AeroBeams.force_scaling — Method

force_scaling(S::Vector{Matrix{Float64}})

Gets the appropriate force scaling for the linear system of equations

Arguments

S::Vector{Matrix{Float64}} = array of elemental sectional compliance matrices

AeroBeams.get_element_distributed_loads — Method

getelementdistributedloads(parent::Beam,Δℓ::Float64,x1n1::Float64)

Gets the distributed loads in the element's local coordinate

Arguments

parent::Beam

Δℓ::Float64

x1_n1::Float64

AeroBeams.get_equivalent_states_rates! — Method

getequivalentstates_rates!(problem::Problem)

Gets the equivalent states' rates at the begin of the current time step

Arguments

problem::Problem

AeroBeams.get_hinged_nodes_matrices — Method

gethingednodes_matrices(parent::Beam,nodesLocalID::Vector{Int64})

Gets the TF matrices resulting from hinged and not hinged nodal DoFs times the identity matrix

Arguments

parent::Beam

nodesLocalID::Vector{Int64}

AeroBeams.get_initial_displacements_derivatives! — Method

getinitialdisplacements_derivatives!(beam::Beam)

Gets the derivatives of the initial generalized displacements

Arguments

beam::Beam

AeroBeams.get_mode_shapes! — Method

getmodeshapes!(problem::Problem,eigenvectorsOscillatory::Matrix{T},frequenciesOscillatory::Vector{Float64},dampingsOscillatory::Vector{Float64})

Gets the mode shapes given by the eigenvectors

Arguments

problem::Problem

eigenvectorsOscillatory::Matrix{T}

frequenciesOscillatory::Vector{Float64}

dampingsOscillatory::Vector{Float64}

AeroBeams.get_rotation_tensors! — Method

getrotationtensors!(beam::Beam)

Validates the input rotation parameters and rotation parametrization, and gets the corresponding rotation tensors from basis A to basis b and of the cross-section

Arguments

beam::Beam

AeroBeams.get_special_nodes! — Method

getspecialnodes!(model::Model)

Gets the special nodes in the system of equations: connection, boundary, and BC'ed nodes

Arguments

model::Model

AeroBeams.get_system_indices! — Method

getsystemindices!(model::Model)

Gets the indices (for equations and DOFs) of the system of equations

Arguments

model::Model

AeroBeams.get_velocity_dofs_to_update! — Method

getvelocitydofstoupdate!(beam::Beam)

Gets the velocity DOFs (V and Ω) to be updated on the initial dynamic solution

Arguments

beam::Beam

AeroBeams.gravitational_loads — Method

gravitational_loads(model::Model,element::Element,σ::Float64)

Computes the nodal resultants from the distributed gravitational loads on the current element

Arguments

model::Model

element::Element

σ::Float64

AeroBeams.gravitational_loads_Jacobians — Method

gravitationalloadsJacobians(element::Element)

Computes the contributions of the gravitational loads to the Jacobian matrix

Arguments

element::Element

AeroBeams.initialize_basis_A_rotation! — Method

initializebasisA_rotation!(model::Model)

Initializes the rotation tensor from basis I (fixed, inertial) to basis A, and its transpose

Arguments

model::Model

AeroBeams.initialize_system_arrays! — Method

initializesystemarrays!(problem::Problem)

Initializes the system arrays to the correct size

Arguments

problem::Problem

AeroBeams.initialize_time_variables! — Method

initializetimevariables!(problem::Problem)

Validates and initializes the time variables

Arguments

problem::Problem

AeroBeams.interpolate_distributed_loads — Method

interpolatedistributedloads(problem::Problem,loadArray::Array{Float64})

Interpolates the loads array at the current time

Arguments

problem::Problem

loadArray::Array{Float64}

AeroBeams.mul3 — Method

mul3(A1::Matrix{<:Number},A2::Matrix{<:Number},A3::Matrix{<:Number},b::Vector{<:Number})

Gets the scalar product of a third-order tensor represented by matrices A1, A2, and A3 with the vector b

Arguments

A1::Matrix{<:Number}

A2::Matrix{<:Number}

A3::Matrix{<:Number}

b::Vector{<:Number}

AeroBeams.plot_undeformed_assembly — Function

plotundeformedassembly(model::Model)

Plots the nodal coordinates of the assembly of beams

Arguments

model::Model

AeroBeams.position_vector_from_curvature — Method

positionvectorfrom_curvature(R0::Matrix{Float64}, k::Vector{Float64}, x1::Float64)

Gets the position vector at an arclength value

Arguments

R0::Matrix{Float64} = initial rotation tensor (at arclength position zero)

k::Vector{Float64} = curvature vector

x1::Float64 = arclength position

AeroBeams.precompute_distributed_loads! — Method

precomputedistributedloads!(problem::Problem)

Pre-computes the distributed loads over all elements' nodes, over every time step

Arguments

problem::Problem

AeroBeams.quaternion_from_rotation_tensor — Method

quaternionfromrotation_tensor(R::Matrix{<:Number})

Computes the quaternion (Euler parameters) given a rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

AeroBeams.rotation_angle — Method

rotation_angle(p::Vector{<:Union{Float64,ComplexF64}})

Gets the rotation angle

Arguments

p::Vector{<:Union{Float64,ComplexF64}} = rotation parameters

AeroBeams.rotation_parameter_scaling — Method

rotationparameterscaling(p::Vector{<:Union{Float64,ComplexF64}})

Scales the Wiener-Milenkovic rotation parameters

Arguments

p::Vector{<:Union{Float64,ComplexF64}} = unscaled rotation parameters

AeroBeams.rotation_parameters_WM — Method

rotationparametersWM(R::Matrix{<:Number})

Computes the Wiener-Milenkovic rotation parameters given a rotation tensor

Arguments

R::Matrix{<:Number} = rotation tensor

AeroBeams.rotation_tensor_E313 — Method

rotationtensorE321(p::::Vector{<:Union{Float64,ComplexF64}})

Gets the rotation tensor according to Euler parameters sequence 3-1-3

Arguments

p::Vector{<:Union{Float64,ComplexF64}} = rotation parameters

AeroBeams.rotation_tensor_E321 — Method

rotationtensorE321(p::Vector{<:Union{Float64,ComplexF64}})

Gets the rotation tensor according to Euler parameters sequence 3-2-1

Arguments

p::Vector{<:Union{Float64,ComplexF64}} = rotation parameters

AeroBeams.rotation_tensor_WM — Method

rotationtensorWM(p::Vector{<:Union{Float64,ComplexF64}})

Gets the rotation tensor according to Wiener-Milenkovic parameters

Arguments

p::Vector{<:Union{Float64,ComplexF64}} = rotation parameters

AeroBeams.rotation_tensor_derivatives_extended_parameters — Method

rotationtensorderivativesextendedparameters(p,pNorm,λ,ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²)

Gets the derivatives of the rotation tensor with respect to the extended rotation parameters

Arguments

p,pNorm,λ,ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²

AeroBeams.rotation_tensor_derivatives_extended_parameters — Method

rotationtensorderivativesextendedparameters(p::Vector{Float64})

Gets the derivatives of the rotation tensor with respect to the extended rotation parameters

Arguments

p::Vector{Float64}

AeroBeams.rotation_tensor_derivatives_scaled_parameters — Method

rotationtensorderivativesscaledparameters(ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²)

Gets the derivatives of the rotation tensor with respect to the scaled rotation parameters

Arguments

ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,υ,υ²

AeroBeams.rotation_tensor_derivatives_time_extended_parameters — Method

rotationtensorderivativestimeextendedparameters(ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,Θps1,Θps2,Θps3,υ,υ²,υ²ps1,υ²ps2,υ²ps3,ps1dot,ps2dot,ps3dot,ps1p1,ps2p1,ps3p1,ps1p2,ps2p2,ps3p2,ps1p3,ps2p3,ps3p3)

Gets the derivatives of the time derivative of the rotation tensor with respect to the extended rotation parameters

Arguments

ps,ps0,ps1,ps2,ps3,ps1s,ps2s,ps3s,ps1ps2,ps1ps3,ps2ps3,Θ,Θps1,Θps2,Θps3,υ,υ²,υ²ps1,υ²ps2,υ²ps3,ps1dot,ps2dot,ps3dot,ps1p1,ps2p1,ps3p1,ps1p2,ps2p2,ps3p2,ps1p3,ps2p3,ps3_p3

AeroBeams.rotation_tensor_from_curvature — Method

rotationtensorfrom_curvature(R0::Matrix{Float64}, k::Vector{Float64}, x1::Float64)

Gets the rotation tensor at an arclength position

Arguments

R0::Matrix{Float64} = initial rotation tensor (at arclength position zero)

k::Vector{Float64} = curvature vector

x1::Float64 = arclength position

AeroBeams.rotation_tensor_time_derivative — Method

rotationtensortimederivative(Rps1,Rps2,Rps3,ps_p,pdot)

Gets the time derivative of the rotation tensor

Arguments

Rps1,Rps2,Rps3,psp,pdot

AeroBeams.rotation_tensor_time_derivative — Method

rotationtensortime_derivative(p::Vector{Float64},pdot::Vector{Float64})

Gets the derivatives of the time derivative of the rotation tensor, given the rotation parameters and their rates

Arguments

p::Vector{Float64}

pdot::Vector{Float64}

AeroBeams.round_off! — Function

round_off!(x)

Rounds the array to input tolerance (defaults to machine epsilon)

Arguments

x

AeroBeams.save_load_factor_data! — Method

saveloadfactor_data!(problem::Problem,σ::Float64,x::Vector{Float64})

Saves the solution at the current load factor

Arguments

problem::Problem

σ::Float64

x::Vector{Float64}

AeroBeams.save_time_step_data! — Method

savetimestep_data!(problem::Problem,timeNow::Number)

Saves the solution at the current time step

Arguments

problem::Problem

AeroBeams.scaling_derivatives_extended_parameters — Method

scalingderivativesextended_parameters(λ::Float64,p::Vector{Float64},pNorm::Float64)

Gets the derivatives of the scaling factor with respect to the extended rotation parameters

Arguments

λ::Float64 = scaling factor

p::Vector{Float64} = rotation parameters

pNorm::Float64 = norm of rotation parameters

AeroBeams.set_BCs! — Method

set_BCs!(model::Model)

Sets the BCs onto the model

Arguments

model::Model

AeroBeams.set_initial_states! — Function

setinitialstates!(problem::Problem,skipSizeAssertion::Bool=false)

Sets the initial elemental and nodal states into the states array

Arguments

problem::Problem

skipSizeAssertion::Bool

AeroBeams.set_nodal_coordinates! — Method

setnodalcoordinates!(beam::Beam)

Sets the nodal coordinates of each element into the beam

Arguments

beam::Beam

AeroBeams.solve! — Method

solve!(problem::Problem)

Solves a problem

Arguments

problem::Problem

AeroBeams.solve_NewtonRaphson! — Method

solve_NewtonRaphson!(problem::Problem)

Solves the nonlinear system of equations at current time step

Arguments

problem::Problem

AeroBeams.solve_dynamic! — Method

solve_dynamic!(problem::Problem)

Solves a dynamic problem

Arguments

problem::Problem

AeroBeams.solve_eigen! — Method

solve_eigen!(problem::Problem)

Solves an eigenproblem

Arguments

problem::Problem

AeroBeams.solve_initial_dynamic! — Method

solveinitialdynamic!(problem::Problem)

Gets consistent initial conditions and solves the first time step

Arguments

problem::Problem

AeroBeams.solve_linear_system! — Method

solvelinearsystem!(problem::Problem)

Solves the linear system of equations at current time step and load factor

Arguments

problem::Problem

AeroBeams.solve_steady! — Method

solve_steady!(problem::Problem)

Solves a steady problem (includes trim and the steady part of eigen problems)

Arguments

problem::Problem

AeroBeams.solve_time_step! — Method

solvetimestep!(problem::Problem)

Solves the current time step

Arguments

problem::Problem

AeroBeams.special_node_arrays! — Method

specialnodearrays!(problem::Problem,model::Model,specialNode::SpecialNode)

Gets the nodal contributions to the system's arrays (residual, Jacobian, inertia)

Arguments

problem::Problem

model::Model

specialNode::SpecialNode

AeroBeams.special_node_follower_loads_jacobians! — Method

specialnodefollowerloadsjacobians!(problem::Problem,specialNode::SpecialNode)

Gets the contributions from the nodal follower loads to the Jacobian matrix

Arguments

problem::Problem

specialNode::SpecialNode

AeroBeams.special_node_jacobian! — Method

specialnodejacobian!(problem::Problem,model::Model,specialNode::SpecialNode)

Computes the contributions from the current node to the Jacobian matrix

Arguments

problem::Problem

model::Model

specialNode::SpecialNode

AeroBeams.special_node_residual! — Method

specialnoderesidual!(problem::Problem,model::Model,specialNode::SpecialNode)

Computes the contributions from the current node to the residual array

Arguments

problem::Problem

model::Model

specialNode::SpecialNode

AeroBeams.special_node_states! — Method

specialnodestates!(problem::Problem,model::Model,specialNode::SpecialNode)

Gets the nodal states

Arguments

problem::Problem

model::Model

specialNode::SpecialNode

AeroBeams.tangent_operator_transpose_WM — Method

tangentoperatortranspose_WM(ps::Vector{Float64},ps0::Float64,υ²::Float64)

Gets the transpose of tangent operator tensor according to Wiener-Milenkovic parameters

Arguments

ps::Vector{Float64} = scaled rotation parameters

ps0::Float64

υ²::Float64

AeroBeams.tangent_operator_transpose_WM — Method

tangentoperatortranspose_WM(p::Vector{Float64})

Gets the transpose of tangent operator tensor according to Wiener-Milenkovic parameters

Arguments

p::Vector{Float64} = rotation parameters

AeroBeams.tangent_operator_transpose_inverse_WM — Method

tangentoperatortransposeinverseWM(p::Vector{<:Union{Float64,ComplexF64}})

Gets the inverse of the transpose of the tangent operator tensor according to Wiener-Milenkovic parameters

Arguments

p::Vector{<:Union{Float64,ComplexF64}} = rotation parameters

AeroBeams.tangent_operator_transpose_inverse_WM — Method

tangentoperatortransposeinverseWM(ps::Vector{Float64},ps0::Float64)

Gets the inverse of the transpose of the tangent operator tensor according to Wiener-Milenkovic parameters

Arguments

ps::Vector{Float64} = scaled rotation parameters

ps0::Float64

AeroBeams.tangent_tensor_functions_derivatives_extended_parameters — Method

tangenttensorfunctionsderivativesextendedparameters(HT,ps1,ps2,ps3,υ²,υ²ps1,υ²ps2,υ²ps3,ps_p)

Gets the derivatives of the tangent tensor's transpose and its inverse with respect to the extended rotation parameters

Arguments

HT,ps1,ps2,ps3,υ²,υ²ps1,υ²ps2,υ²ps3,psp

AeroBeams.tangent_tensor_transpose_derivatives_extended_parameters — Method

tangenttensortransposederivativesextended_parameters(p::Vector{Float64})

Gets the derivatives of the tangent tensor's transpose with respect to the extended rotation parameters

Arguments

p::Vector{Float64}

AeroBeams.tilde — Method

tilde(v::Vector{<:Number})

Gets the skew-symmetric matrix associated with a vector

Arguments

v::Vector{<:Number} = three-element vector

AeroBeams.time_march! — Method

time_march!(problem::Problem)

Marches the dynamic problem in time

Arguments

problem::Problem

AeroBeams.update_BC_data! — Function

updateBCdata!(bc::BC,timeNow::Float64=0.0)

Updates the boundary conditions at the current time

Arguments

bc::BC

timeNow::Float64

AeroBeams.update_basis_A_orientation! — Method

updatebasisA_orientation!(problem::Problem)

Updates the orientation of the basis A for the next time step

Arguments

problem::Problem

AeroBeams.update_beam! — Method

update_beam!(beam::Beam)

Validates and updates the beam construction

Arguments

beam::Beam

AeroBeams.update_element_distributed_loads! — Method

updateelementdistributed_loads!(element::Element,loadType::String,loadFun::Function)

Update the distributed loads in the element's local coordinate

Arguments

element::Element

loadType::String

loadFun::Function

AeroBeams.update_initial_conditions! — Method

updateinitialconditions!(model::Model)

Updates the initial condition states on all elements of the assembly

Arguments

model::Model

AeroBeams.update_initial_velocities! — Method

updateinitialvelocities!(problem::Problem)

Updates the velocity states by running a very small time step

Arguments

problem::Problem

AeroBeams.update_loads_trim_links_global_ids! — Method

updateloadstrimlinksglobal_ids!(model::Model)

Updates the nodes' global IDs of the loads trim links

Arguments

model::Model

AeroBeams.update_model! — Method

update_model!(model::Model)

Updates the model with its current settings

Arguments

model::Model

AeroBeams.update_special_node_jacobian! — Function

updatespecialnodejacobian!(jacobian,forceScaling,Fp,Mp,ζonElements,eqsFu,eqsFp,eqsFF,eqsFM,eqsFFsep,eqsFMsep,DOFuF,DOFpM,DOFtrimLoads,uIsPrescribed,pIsPrescribed,isLoad=trues(6),isTrim=falses(6))

Updates the Jacobian matrix with the contributions from the current BC (if any) on the current node

Arguments

problem::Problem

model::Model

specialNode::SpecialNode

AeroBeams.update_states! — Method

update_states!(problem::Problem)

Updates the elemental and nodal states

Arguments

problem::Problem

AeroBeams.update_time_variables! — Method

updatetimevariables!(problem::Problem,timeIndex::Int64)

Updates the time variables (time, time step, time indices)

Arguments

problem::Problem

timeIndex::Int64

AeroBeams.validate_and_update_motion_basis_A! — Method

validateandupdatemotionbasis_A!(model::Model)

Validates and updates the motion variables of basis A

Arguments

model::Model

AeroBeams.validate_beam! — Method

validate_beam!(beam::Beam)

Validates the beam inputs

Arguments

beam::Beam

AeroBeams.validate_connected_beams — Method

validateconnectedbeams(beam::Beam)

Checks that the beam connections are consistent

Arguments

beam::Beam

AeroBeams.validate_distributed_loads! — Method

validatedistributedloads!(beam::Beam)

Validates and updates the distributed loads

Arguments

beam::Beam

AeroBeams.validate_hinged_nodes! — Method

validatehingednodes!(beam::Beam)

Validates the hinged nodes and DoFs

Arguments

beam::Beam

AeroBeams.validate_initial_conditions! — Method

validateinitialconditions!(beam::Beam)

Validate initial generalized displacements/velocities defined in basis b, and transform to function if input as constant vector

Arguments

beam::Beam

AeroBeams.validate_model! — Method

validate_model!(model::Model)

Validates the inputs to the model

Arguments

model::Model

AeroBeams.validate_normalized_nodal_positions! — Method

validatenormalizednodal_positions!(beam::Beam)

Validates the normalized nodal positions if they were input, or updates them if they were not

Arguments

beam::Beam

AeroBeams.validate_rotation_parametrization — Method

validaterotationparametrization(beam::Beam)

Validates the input rotation parameters and rotation parametrization

Arguments

beam::Beam

AeroBeams.validate_sectional_matrices — Method

validatesectionalmatrices(beam::Beam)

Checks that sectional matrices are input as a single one for the whole beam or are input in a per element basis

Arguments

beam::Beam

AeroBeams.validate_unit_system — Method

validateunitsystem(units::UnitSystem)

Validates the unit system

Fields:

units::UnitSystem